### 2022-04-20 #4330 @editorialbot
[PRE REVIEW]: egobox, a Rust toolbox for efficient global optimization
**Submitting author:** @relf (<a href="http://orcid.org/0000-0001-5479-2961">Rémi Lafage</a>)
**Repository:** https://github.com/relf/egobox
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.2.1
