### 2022-04-20 #4330 @arfon
@relf - please format the affiliations as shown in the example paper in our docs: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
