### 2022-04-21 dda7719
Dockerfile to compile LLVM on manylinux2014
