### 2022-04-21 #4330 @relf
> @relf - please format the affiliations as shown in the example paper in our docs: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

@arfon, sorry for that. I've used the docker command and did not see the glitch as the pdf was generated. It is fixed.
