### 2022-04-21 9bcf4ac
Add CC and CXX required for cobyla build
