### 2022-04-21 5df68da
Make serialization optional in gp (#16)
* Make serialization optional in gp

* Update doc
