+++ moe/src/algorithm.rs
@@ -3,7 +3,7 @@ use crate::errors::MoeError;
 use crate::errors::Result;
 use crate::expertise_macros::*;
 use crate::parameters::{
-    CorrelationSpec, MoeFit, MoeParams, MoePredict, Recombination, RegressionSpec,
+    CorrelationSpec, Expert, MoeFit, MoeParams, Recombination, RegressionSpec,
 };
 use crate::surrogates::*;
 use log::{debug, info, trace};
@@ -41,9 +41,8 @@ macro_rules! check_allowed {
 }
 
 impl<R: Rng + SeedableRng + Clone> MoeFit for MoeParams<f64, R> {
-    fn fit_for_predict(&self, xt: &Array2<f64>, yt: &Array2<f64>) -> Result<Box<dyn MoePredict>> {
-        self.fit(xt, yt)
-            .map(|moe| Box::new(moe) as Box<dyn MoePredict>)
+    fn fit_for_predict(&self, xt: &Array2<f64>, yt: &Array2<f64>) -> Result<Box<dyn Expert>> {
+        self.fit(xt, yt).map(|moe| Box::new(moe) as Box<dyn Expert>)
     }
 }
 
@@ -131,7 +130,7 @@ impl<R: Rng + SeedableRng + Clone> MoeParams<f64, R> {
 
     /// Select the surrogate which gives the smallest prediction error on the given data
     /// The error is computed using cross-validation
-    fn find_best_expert(&self, nx: usize, data: &Array2<f64>) -> Result<Box<dyn GpSurrogate>> {
+    fn find_best_expert(&self, nx: usize, data: &Array2<f64>) -> Result<Box<dyn Surrogate>> {
         let xtrain = data.slice(s![.., ..nx]).to_owned();
         let ytrain = data.slice(s![.., nx..]).to_owned();
         let mut dataset = Dataset::from((xtrain.clone(), ytrain.clone()));
@@ -172,34 +171,32 @@ impl<R: Rng + SeedableRng + Clone> MoeParams<f64, R> {
                 .unwrap();
             (map_error[argmin].0.clone(), Some(map_error[argmin].1))
         };
-        let best_expert_params: std::result::Result<Box<dyn GpSurrogateParams>, MoeError> =
-            match best.0.as_str() {
-                "Constant_SquaredExponential" => {
-                    Ok(make_surrogate_params!(Constant, SquaredExponential))
-                }
-                "Constant_AbsoluteExponential" => {
-                    Ok(make_surrogate_params!(Constant, AbsoluteExponential))
-                }
-                "Constant_Matern32" => Ok(make_surrogate_params!(Constant, Matern32)),
-                "Constant_Matern52" => Ok(make_surrogate_params!(Constant, Matern52)),
-                "Linear_SquaredExponential" => {
-                    Ok(make_surrogate_params!(Linear, SquaredExponential))
-                }
-                "Linear_AbsoluteExponential" => {
-                    Ok(make_surrogate_params!(Linear, AbsoluteExponential))
-                }
-                "Linear_Matern32" => Ok(make_surrogate_params!(Linear, Matern32)),
-                "Linear_Matern52" => Ok(make_surrogate_params!(Linear, Matern52)),
-                "Quadratic_SquaredExponential" => {
-                    Ok(make_surrogate_params!(Quadratic, SquaredExponential))
-                }
-                "Quadratic_AbsoluteExponential" => {
-                    Ok(make_surrogate_params!(Quadratic, AbsoluteExponential))
-                }
-                "Quadratic_Matern32" => Ok(make_surrogate_params!(Quadratic, Matern32)),
-                "Quadratic_Matern52" => Ok(make_surrogate_params!(Quadratic, Matern52)),
[…]
