+++ moe/src/parameters.rs
@@ -1,4 +1,4 @@
-use crate::errors::Result;
+use crate::errors::{MoeError, Result};
 use crate::surrogates::Surrogate;
 use bitflags::bitflags;
 #[allow(unused_imports)]
@@ -7,7 +7,7 @@ use egobox_gp::correlation_models::{
 };
 #[allow(unused_imports)]
 use egobox_gp::mean_models::{ConstantMean, LinearMean, QuadraticMean};
-use linfa::Float;
+use linfa::{Float, ParamGuard};
 use linfa_clustering::GaussianMixtureModel;
 use ndarray::Array2;
 use ndarray_rand::rand::{Rng, SeedableRng};
@@ -104,7 +104,7 @@ pub trait MoeFit {
 
 /// Mixture of experts parameters
 #[derive(Clone)]
-pub struct MoeParams<F: Float, R: Rng + Clone> {
+pub struct MoeValidParams<F: Float, R: Rng + Clone> {
     /// Number of clusters (i.e. number of experts)
     n_clusters: usize,
     /// [Recombination] mode
@@ -122,43 +122,21 @@ pub struct MoeParams<F: Float, R: Rng + Clone> {
     rng: R,
 }
 
-impl<F: Float> MoeParams<F, Isaac64Rng> {
-    /// Constructor of Moe parameters with `n_clusters`.
-    ///
-    /// Default values are provided as follows:
-    ///
-    /// * recombination: `Smooth`
-    /// * regression_spec: `ALL`
-    /// * correlation_spec: `ALL`
-    /// * kpls_dim: `None`
-    #[allow(clippy::new_ret_no_self)]
-    pub fn new(n_clusters: usize) -> MoeParams<F, Isaac64Rng> {
-        Self::new_with_rng(n_clusters, Isaac64Rng::from_entropy())
-    }
-}
-
-impl<F: Float> Default for MoeParams<F, Isaac64Rng> {
-    fn default() -> MoeParams<F, Isaac64Rng> {
-        MoeParams::new(1)
-    }
-}
-
-impl<F: Float, R: Rng + Clone> MoeParams<F, R> {
-    /// Constructor of Moe parameters specifying randon number generator for reproducibility
-    ///
-    /// See [MoeParams::new] for default parameters.
-    pub fn new_with_rng(n_clusters: usize, rng: R) -> MoeParams<F, R> {
-        MoeParams {
-            n_clusters,
+impl<F: Float> Default for MoeValidParams<F, Isaac64Rng> {
+    fn default() -> MoeValidParams<F, Isaac64Rng> {
+        MoeValidParams {
+            n_clusters: 1,
             recombination: Recombination::Smooth(Some(F::one())),
             regression_spec: RegressionSpec::ALL,
             correlation_spec: CorrelationSpec::ALL,
             kpls_dim: None,
             gmm: None,
-            rng,
+            rng: Isaac64Rng::from_entropy(),
         }
     }
+}
 
+impl<F: Float, R: Rng + Clone> MoeValidParams<F, R> {
     /// The number of clusters, hence the number of experts of the mixture.
     pub fn n_clusters(&self) -> usize {
         self.n_clusters
@@ -184,20 +162,66 @@ impl<F: Float, R: Rng + Clone> MoeParams<F, R> {
         self.kpls_dim
     }
 
+    /// The optional number of PLS components
+    pub fn gmm(&self) -> &Option<Box<GaussianMixtureModel<F>>> {
+        &self.gmm
+    }
+
     /// The random generator
     pub fn rng(&self) -> R {
         self.rng.clone()
     }
+}
+
+#[derive(Clone)]
[…]
