+++ moe/examples/norm1.rs
@@ -1,5 +1,6 @@
 use egobox_doe::{Lhs, SamplingMethod};
 use egobox_moe::{Expert, Moe, Recombination};
+use linfa::ParamGuard;
 use ndarray::{arr2, Array2, Axis};
 use std::error::Error;
 
@@ -10,9 +11,10 @@ fn norm1(x: &Array2<f64>) -> Array2<f64> {
 fn main() -> Result<(), Box<dyn Error>> {
     let xtrain = Lhs::new(&arr2(&[[-1., 1.], [-1., 1.]])).sample(200);
     let ytrain = norm1(&xtrain);
-    let moe1 = Moe::params(1).fit(&xtrain, &ytrain)?;
+    let moe1 = Moe::params(1).check_unwrap().fit(&xtrain, &ytrain)?;
     let moe5 = Moe::params(6)
         .set_recombination(Recombination::Hard)
+        .check_unwrap()
         .fit(&xtrain, &ytrain)?;
 
     let xtest = Lhs::new(&arr2(&[[-1., 1.], [-1., 1.]])).sample(50);

[…]
