+++ moe/examples/moe_norm1.rs
@@ -1,6 +1,7 @@
 use csv::ReaderBuilder;
 use egobox_doe::{FullFactorial, SamplingMethod};
-use egobox_moe::{Expert, Moe};
+use egobox_moe::Moe;
+use linfa::{traits::Fit, Dataset};
 use ndarray::{arr2, s, Array2, Axis};
 use ndarray_csv::Array2Reader;
 use ndarray_npy::write_npy;
@@ -20,9 +21,10 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     let data_train: Array2<f64> = reader.deserialize_array2((200, 3))?;
 
-    let xtrain = data_train.slice(s![.., ..2]).to_owned();
-    let ytrain = data_train.slice(s![.., 2..]).to_owned();
-    let moe = Moe::params(4).fit(&xtrain, &ytrain)?;
+    let xtrain = data_train.slice(s![.., ..2_usize]).to_owned();
+    let ytrain = data_train.slice(s![.., 2_usize..]).to_owned();
+    let ds = Dataset::new(xtrain, ytrain);
+    let moe = Moe::params(4).fit(&ds)?;
 
     let xlimits = arr2(&[[-1., 1.], [-1., 1.]]);
     let xtest = FullFactorial::new(&xlimits).sample(100);
