+++ moe/examples/clustering.rs
@@ -1,5 +1,6 @@
 use egobox_doe::{Lhs, SamplingMethod};
-use egobox_moe::{Expert, Moe, Recombination};
+use egobox_moe::{Moe, Recombination};
+use linfa::prelude::{Dataset, Fit};
 use ndarray::{arr2, Array, Array2, Axis, Zip};
 use std::error::Error;
 
@@ -20,10 +21,11 @@ fn f3parts(x: &Array2<f64>) -> Array2<f64> {
 fn main() -> Result<(), Box<dyn Error>> {
     let xtrain = Lhs::new(&arr2(&[[0., 1.]])).sample(50);
     let ytrain = f3parts(&xtrain);
-    let moe1 = Moe::params(1).fit(&xtrain, &ytrain)?;
+    let ds = Dataset::new(xtrain, ytrain);
+    let moe1 = Moe::params(1).fit(&ds)?;
     let moe3 = Moe::params(3)
         .set_recombination(Recombination::Hard)
-        .fit(&xtrain, &ytrain)?;
+        .fit(&ds)?;
 
     let xtest = Array::linspace(0., 1., 101).insert_axis(Axis(1));
     let ytest = f3parts(&xtest);
