+++ moe/src/clustering.rs
@@ -1,7 +1,7 @@
 #![allow(dead_code)]
 use crate::algorithm::{sort_by_cluster, Moe};
 use crate::gaussian_mixture::GaussianMixture;
-use crate::parameters::{CorrelationSpec, Expert, Recombination, RegressionSpec};
+use crate::parameters::{CorrelationSpec, Recombination, RegressionSpec};
 use log::debug;
 
 use linfa::dataset::{Dataset, DatasetView};
@@ -103,7 +103,7 @@ pub fn find_best_number_of_clusters<R: Rng + SeedableRng + Clone>(
                     .set_correlation_spec(correlation_spec)
                     //.set_kpls_dim(Some(1))
                     .set_gmm(Some(gmm.clone()))
-                    .fit(train.records(), train.targets())
+                    .fit(&train)
                 {
                     let xytrain =
                         concatenate(Axis(1), &[train.records().view(), train.targets.view()])
@@ -129,30 +129,28 @@ pub fn find_best_number_of_clusters<R: Rng + SeedableRng + Clone>(
                     }
                     let actual = valid.targets();
                     let mixture = mixture.set_recombination(Recombination::Hard);
-                    let h_error =
-                        if let Ok(pred) = mixture.predict_values(&valid.records().to_owned()) {
-                            // write_npy(format!("valid_x_{}_{}.npy", n_clusters, k), valid.records())
-                            //     .expect("valid x saved");
-                            // write_npy(format!("valid_y_{}_{}.npy", n_clusters, k), actual)
-                            //     .expect("valid y saved");
-                            // write_npy(format!("pred_{}_{}.npy", n_clusters, k), &pred)
-                            //     .expect("pred saved");
-                            pred.sub(actual).mapv(|x| x * x).sum().sqrt()
-                                / actual.mapv(|x| x * x).sum().sqrt()
-                        } else {
-                            ok = false;
-                            1.0
-                        };
+                    let h_error = if let Ok(pred) = mixture.predict_values(valid.records()) {
+                        // write_npy(format!("valid_x_{}_{}.npy", n_clusters, k), valid.records())
+                        //     .expect("valid x saved");
+                        // write_npy(format!("valid_y_{}_{}.npy", n_clusters, k), actual)
+                        //     .expect("valid y saved");
+                        // write_npy(format!("pred_{}_{}.npy", n_clusters, k), &pred)
+                        //     .expect("pred saved");
+                        pred.sub(actual).mapv(|x| x * x).sum().sqrt()
+                            / actual.mapv(|x| x * x).sum().sqrt()
+                    } else {
+                        ok = false;
+                        1.0
+                    };
                     h_errors.push(h_error);
                     let mixture = mixture.set_recombination(Recombination::Smooth(None));
-                    let s_error =
-                        if let Ok(pred) = mixture.predict_values(&valid.records().to_owned()) {
-                            pred.sub(actual).mapv(|x| x * x).sum().sqrt()
-                                / actual.mapv(|x| x * x).sum().sqrt()
-                        } else {
-                            ok = false;
-                            1.0
-                        };
+                    let s_error = if let Ok(pred) = mixture.predict_values(valid.records()) {
+                        pred.sub(actual).mapv(|x| x * x).sum().sqrt()
+                            / actual.mapv(|x| x * x).sum().sqrt()
+                    } else {
+                        ok = false;
+                        1.0
+                    };
[…]
