+++ moe/examples/norm1.rs
@@ -13,9 +13,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     let ytrain = norm1(&xtrain);
     let ds = Dataset::new(xtrain, ytrain);
     let moe1 = Moe::params(1).fit(&ds)?;
-    let moe5 = Moe::params(6)
-        .set_recombination(Recombination::Hard)
-        .fit(&ds)?;
+    let moe5 = Moe::params(6).recombination(Recombination::Hard).fit(&ds)?;
 
     let xtest = Lhs::new(&arr2(&[[-1., 1.], [-1., 1.]])).sample(50);
     let ytest = norm1(&xtest);

[…]
