### 2022-04-24 0e94d6e
MoE improvements (#15)
* Improve doc

* Implement moe save/load (feature persistent)

* Refactor MoE traits

* Rename GpSurrogate to Surrogate
* Rename MoePredict to Expert
* Implement Surrogate as Expert

* Rename fit_for_predict

* Implement ParamGuard for MoeParams

* Implement MoeFit and fit or MoeParams

* Remove Expert trait, Implement Fit for MoeParams

* Rename MoeParams setters
