+++ moe/src/errors.rs
@@ -1,7 +1,4 @@
-// use egobox_gp::GpError;
-// use ndarray_linalg::error::LinalgError;
 use thiserror::Error;
-// use std::fmt::{self, Display};
 
 /// A result type for Moe algorithm
 pub type Result<T> = std::result::Result<T, MoeError>;
@@ -25,6 +22,7 @@ pub enum MoeError {
     #[error("Clustering error: {0}")]
     ClusteringError(String),
     /// When error during saving
+    #[cfg(feature = "persistent")]
     #[error("Save error: {0}")]
     SaveError(#[from] serde_json::Error),
     /// When error during loading
@@ -33,4 +31,10 @@ pub enum MoeError {
     /// When error during loading
     #[error("Load error: {0}")]
     LoadError(String),
+    /// When error during loading
+    #[error("InvalidValue error: {0}")]
+    InvalidValueError(String),
+    /// When a linfa error occurs
+    #[error(transparent)]
+    LinfaError(#[from] linfa::error::Error),
 }
