+++ joss/paper.bib
@@ -16,8 +16,9 @@
   volume={53},
   number={5},
   pages={935--952},
+  doi={10.1007/s00158-015-1395-9},
   year={2016},
-  publisher={Springer}
+  publisher={Springer},
 }
 
 @inproceedings{Bartoli2016,
@@ -25,7 +26,8 @@
   author={Bartoli, Nathalie and Bouhlel, Mohamed-Amine and Kurek, Igor and Lafage, R{\'e}mi and Lefebvre, Thierry and Morlier, Joseph and Priem, R{\'e}my and Stilz, Vivien and Regis, Rommel},
   booktitle={17th AIAA/ISSMO Multidisciplinary analysis and optimization conference},
   pages={4001},
-  year={2016}
+  doi={10.2514/6.2016-4001},
+  year={2016},
 }
 
 @article{Bartoli2019,
@@ -33,9 +35,9 @@
   journal = {Aerospace Science and Technology},
   volume = {90},
   pages = {85-102},
+  doi = {10.1016/j.ast.2019.03.041},
   year = {2019},
   issn = {1270-9638},
-  doi = {https://doi.org/10.1016/j.ast.2019.03.041},
   url = {https://www.sciencedirect.com/science/article/pii/S1270963818306011},
   author = {N. Bartoli and T. Lefebvre and S. Dubreuil and R. Olivanti and R. Priem and N. Bons and J.R.R.A. Martins and J. Morlier},
   keywords = {Surrogate modeling, Global optimization, Multimodal optimization, Mixture of experts, Aerodynamic shape optimization, Wing design},
@@ -49,6 +51,7 @@
   volume={62},
   number={4},
   pages={1739--1765},
+  doi={10.1007/s00158-020-02514-6},
   year={2020},
   publisher={Springer}
 }
@@ -60,7 +63,7 @@
   pages = {102662},
   year = {2019},
   issn = {0965-9978},
-  doi = {https://doi.org/10.1016/j.advengsoft.2019.03.005},
+  doi = {10.1016/j.advengsoft.2019.03.005},
   year = {2019}
 }
 
