### 2022-04-30 #4330 @relf
> Also, your statement of need says why this code is in RUST but it doesn't actually say why this code is needed - e.g., who will use it and for what.

@danielskatz , thanks for your comment. I've (hopefully) clarified this point.
