### 2022-05-02 2388823
Refactor Egor/Moe relation
* Move MoeFit as SurrogateBuilder from moe to ego
* Implement SurrogateBuilder for Moe in ego
* Moe uses linfa Fit interface
