+++ ego/src/egor.rs
@@ -97,9 +97,10 @@ use crate::utils::update_data;
 use crate::utils::{compute_cstr_scales, compute_obj_scale, compute_wb2s_scale};
 use crate::utils::{ei, wb2s};
 use egobox_doe::{Lhs, LhsKind, SamplingMethod};
-use egobox_moe::{CorrelationSpec, Moe, MoeFit, RegressionSpec, Surrogate};
+use egobox_moe::{CorrelationSpec, Moe, MoeParams, RegressionSpec, Surrogate};
 use env_logger::{Builder, Env};
 use finitediff::FiniteDiff;
+use linfa::{traits::Fit, Dataset, ParamGuard};
 use log::{debug, info};
 use ndarray::{concatenate, s, Array, Array1, Array2, ArrayBase, Axis, Data, Ix1, Ix2, Zip};
 use ndarray_linalg::Scalar;
@@ -112,22 +113,19 @@ use rand_isaac::Isaac64Rng;
 const DOE_INITIAL_FILE: &str = "egor_initial_doe.npy";
 const DOE_FILE: &str = "egobox_doe.npy";
 
-#[derive(Clone, Copy, Debug)]
-pub struct ApproxValue {
-    pub value: f64,
-    pub tolerance: f64,
-}
-
 /// Data used by internal infill criteria to be optimized using NlOpt
-pub struct ObjData<F> {
+struct ObjData<F> {
     pub scale_obj: F,
     pub scale_cstr: Array1<F>,
     pub scale_wb2: F,
 }
 
-/// An interface for "function under optimization" evaluation
-pub trait Evaluator {
-    fn eval(&self, x: &Array2<f64>) -> Array2<f64>;
+impl<R: Rng + SeedableRng + Clone> SurrogateBuilder for MoeParams<f64, R> {
+    fn train(&self, xt: &Array2<f64>, yt: &Array2<f64>) -> Result<Box<dyn Surrogate>> {
+        let checked = self.check_ref()?;
+        let moe = checked.fit(&Dataset::new(xt.to_owned(), yt.to_owned()))?;
+        Ok(moe).map(|moe| Box::new(moe) as Box<dyn Surrogate>)
+    }
 }
 
 /// EGO optimization parameterization
@@ -173,10 +171,13 @@ pub struct Egor<'a, O: GroupFunc, R: Rng> {
     pub outdir: Option<String>,
     /// If true use <outdir> to retrieve and start from previous results
     pub hot_start: bool,
-    /// MoE parameters (see [egobox_moe])
+    /// An optional surrogate builder used to model objective and constraint
+    /// functions, otherwise [mixture of expert](egobox_moe) is used
     /// Note: if specified takes precedence over individual settings
-    pub moe_params: Option<&'a dyn MoeFit>,
-    /// An evaluator used to run the function under optimization
+    pub surrogate_builder: Option<&'a dyn SurrogateBuilder>,
+    /// An optional evaluator used to run the function under optimization
+    /// allowing to adjust pre or post processing around function evaluation,
+    /// the function is run directly otherwise
     pub evaluator: Option<&'a dyn Evaluator>,
     /// The function under optimization f(x) = [objective, cstr1, ..., cstrn], (n_cstr+1 size)
     pub obj: O,
@@ -216,7 +217,7 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
             expected: None,
             outdir: None,
             hot_start: false,
-            moe_params: None,
+            surrogate_builder: None,
             evaluator: None,
             obj: f,
             rng,
@@ -308,8 +309,11 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
[…]
