### 2022-05-03 ea950f3
Refactor ego/moe relation (#17)
* Improve GP doc

* Move MoeFit as SurrogateBuilder from moe to ego
* Implement SurrogateBuilder for Moe in ego
* Moe uses linfa Fit interface

* Rename Evaluator as PreProcessor, update doc

* Fix doc test
