### 2022-05-05 cc1773a
Rename n_parallel to q_parallel (qEI stategy)
