### 2022-05-05 a23f762
Refactor MixintEgor (#18)
* Rename PreProcessor::eval to run

* Implement Fit for MixintMoeParams. Use Dataset.

* Rename SurrogateParams to MoeBuilder

* Relax test tolerance

* Update doc

* Rename n_parallel to q_parallel (qEI stategy)

* Fix typo
