+++ ego/src/egor.rs
@@ -137,7 +137,7 @@ pub struct Egor<'a, O: GroupFunc, R: Rng> {
     /// Number of starts for multistart approach used for hyperparameters optimization
     pub n_start: usize,
     /// Number of parallel points evaluated for each "function evaluation"
-    pub n_parallel: usize,
+    pub q_parallel: usize,
     /// Number of initial doe drawn using Latin hypercube sampling
     /// Note: n_doe > 0; otherwise n_doe = max(xdim+1, 5)
     pub n_doe: usize,
@@ -151,7 +151,7 @@ pub struct Egor<'a, O: GroupFunc, R: Rng> {
     pub doe: Option<Array2<f64>>,
     /// Matrix (nx, 2) of [lower bound, upper bound] of the nx components of x
     pub xlimits: Array2<f64>,
-    /// Parallel strategy used to define several points (n_parallel) evaluations at each iteration
+    /// Parallel strategy used to define several points (q_parallel) evaluations at each iteration
     pub q_ei: QEiStrategy,
     /// Criterium to select next point to evaluate
     pub infill: InfillStrategy,
@@ -175,8 +175,8 @@ pub struct Egor<'a, O: GroupFunc, R: Rng> {
     /// functions, otherwise [mixture of expert](egobox_moe) is used
     /// Note: if specified takes precedence over individual settings
     pub surrogate_builder: Option<&'a dyn SurrogateBuilder>,
-    /// An optional pre-processor used to preprocess continuous input given
-    /// to the function under optimization, specially with mixed integer
+    /// An optional pre-processor to preprocess continuous input given
+    /// to the function under optimization, specially used with mixed integer
     /// function optimization
     pub pre_proc: Option<&'a dyn PreProcessor>,
     /// The function under optimization f(x) = [objective, cstr1, ..., cstrn], (n_cstr+1 size)
@@ -187,12 +187,21 @@ pub struct Egor<'a, O: GroupFunc, R: Rng> {
 }
 
 impl<'a, O: GroupFunc> Egor<'a, O, Isaac64Rng> {
+    /// Constructor of the optimization of the function `f`
+    ///
+    /// The function `f` shoud return an objective but also constraint values if any.
+    /// Design space is specified by the 2D array `xlimits` which is `[nx, 2]`-shaped and
+    /// constains lower and upper bounds of `x` components.
     pub fn new(f: O, xlimits: &ArrayBase<impl Data<Elem = f64>, Ix2>) -> Egor<'a, O, Isaac64Rng> {
         Self::new_with_rng(f, xlimits, Isaac64Rng::from_entropy())
     }
 }
 
 impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
+    /// Constructor of the optimization of the function `f` with specified random generator
+    /// to get reproducibility.
+    ///
+    /// See [`Egor::new()`]
     pub fn new_with_rng(f: O, xlimits: &ArrayBase<impl Data<Elem = f64>, Ix2>, rng: R) -> Self {
         let env = Env::new().filter_or("EGOBOX_LOG", "info");
         let mut builder = Builder::from_env(env);
@@ -201,7 +210,7 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
         Egor {
             n_eval: 20,
             n_start: 20,
-            n_parallel: 1,
+            q_parallel: 1,
             n_doe: 0,
             n_cstr: 0,
             cstr_tol: 1e-6,
@@ -224,91 +233,122 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
         }
     }
 
[…]
