+++ ego/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "egobox-ego"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["Rémi Lafage <remi.lafage@onera.fr>"]
 edition = "2018"
 description = "A library for efficient global optimization"
@@ -15,9 +15,9 @@ default = []
 persistent = ["serde", "typetag", "egobox-moe/persistent", "serde_json"]
 
 [dependencies]
-egobox-doe = { version = "0.2.1", path="../doe" }
-egobox-gp = { version = "0.2.1", path="../gp" }
-egobox-moe = { version = "0.2.1", path="../moe" }
+egobox-doe = { version = "0.3.0", path="../doe" }
+egobox-gp = { version = "0.3.0", path="../gp" }
+egobox-moe = { version = "0.3.0", path="../moe" }
 
 linfa = { version = "0.5.1", default-features = false }
 linfa-pls = { version = "0.5.1", default-features = false }
