+++ moe/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "egobox-moe"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["Rémi Lafage <remi.lafage@onera.fr>"]
 edition = "2018"
 description = "A library for mixture of expert gaussian processes"
@@ -15,8 +15,8 @@ default = []
 persistent = ["serde", "typetag", "egobox-gp/serializable", "serde_json"]
 
 [dependencies]
-egobox-doe = { version = "0.2.1", path = "../doe" }
-egobox-gp = { version = "0.2.1", path = "../gp" }
+egobox-doe = { version = "0.3.0", path = "../doe" }
+egobox-gp = { version = "0.3.0", path = "../gp" }
 
 linfa = { version = "0.5.1", default-features = false }
 linfa-clustering = { version = "0.5.1", default-features = false }
