### 2022-05-06 e2374b2
Release 0.3.0 (#19)
* Add CHANGELOG

* Bump 0.3.0
