+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "egobox"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["Rémi Lafage <remi.lafage@onera.fr>"]
 edition = "2018"
 description = "A toolbox for efficient global optimization"
@@ -33,10 +33,10 @@ persistent-moe = ["egobox-moe/persistent"]
 persistent-ego = ["egobox-ego/persistent"]
 
 [dependencies]
-egobox-doe = { version = "0.2.1", path="./doe" }
-egobox-gp = { version = "0.2.1", path="./gp" }
-egobox-moe = { version = "0.2.1", path="./moe" }
-egobox-ego = { version = "0.2.1", path="./ego" }
+egobox-doe = { version = "0.3.0", path="./doe" }
+egobox-gp = { version = "0.3.0", path="./gp" }
+egobox-moe = { version = "0.3.0", path="./moe" }
+egobox-ego = { version = "0.3.0", path="./ego" }
 
 linfa = { version = "0.5.1", default-features = false }
 linfa-pls = { version = "0.5.1", default-features = false }
