### 2022-05-06 62bb77d
Add architecture figure
