### 2022-05-07 5a92692
egobox on linux (#20)
* Dockerfile to compile LLVM on manylinux2014

* Add CC and CXX required for cobyla build
