### 2022-05-09 bf0388c
Add LHS optimisation in case of repeated point rejections
