### 2022-05-11 eaa6adc
Adjust optim rosenbrock test
