### 2022-05-13 50ced92
Ego: Add LHS optimization (#21)
* Add LHS optimisation in case of repeated point rejections

* Improve log

* Use classic LHS for random optimization

* Adjust optim rosenbrock test
