+++ ego/src/egor.rs
@@ -91,11 +91,13 @@
 //! ```
 //!
 use crate::errors::{EgoError, Result};
+use crate::lhs_optimizer::LhsOptimizer;
 use crate::sort_axis::*;
 use crate::types::*;
 use crate::utils::update_data;
 use crate::utils::{compute_cstr_scales, compute_obj_scale, compute_wb2s_scale};
 use crate::utils::{ei, wb2s};
+
 use egobox_doe::{Lhs, LhsKind, SamplingMethod};
 use egobox_moe::{CorrelationSpec, Moe, MoeParams, RegressionSpec, Surrogate};
 use env_logger::{Builder, Env};
@@ -112,13 +114,6 @@ use rand_isaac::Isaac64Rng;
 
 const DOE_INITIAL_FILE: &str = "egor_initial_doe.npy";
 const DOE_FILE: &str = "egobox_doe.npy";
-
-/// Data used by internal infill criteria to be optimized using NlOpt
-struct ObjData<F> {
-    pub scale_obj: F,
-    pub scale_cstr: Array1<F>,
-    pub scale_wb2: F,
-}
 
 impl<R: Rng + SeedableRng + Clone> SurrogateBuilder for MoeParams<f64, R> {
     fn train(&self, xt: &Array2<f64>, yt: &Array2<f64>) -> Result<Box<dyn Surrogate>> {
@@ -400,7 +395,7 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
     pub fn suggest(&self, x_data: &Array2<f64>, y_data: &Array2<f64>) -> Array2<f64> {
         let rng = self.rng.clone();
         let sampling = Lhs::new(&self.xlimits).with_rng(rng).kind(LhsKind::Maximin);
-        let (x_dat, _) = self.next_points(0, x_data, y_data, &sampling);
+        let (x_dat, _) = self.next_points(0, x_data, y_data, &sampling, false);
         x_dat
     }
 
@@ -465,10 +460,11 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
             write_npy(filepath, &doe).expect("Write initial doe");
         }
 
-        const MAX_RETRY: i32 = 10;
+        const MAX_RETRY: i32 = 3;
         let mut no_point_added_retries = MAX_RETRY;
+        let mut lhs_optim = false;
         for i in 1..=n_iter {
-            let (x_dat, y_dat) = self.next_points(i, &x_data, &y_data, &sampling);
+            let (x_dat, y_dat) = self.next_points(i, &x_data, &y_data, &sampling, lhs_optim);
             debug!("Try adding {}", x_dat);
             let rejected_count = update_data(&mut x_data, &mut y_data, &x_dat, &y_dat);
             if rejected_count > 0 {
@@ -484,45 +480,50 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
                 no_point_added_retries -= 1;
                 if no_point_added_retries == 0 {
                     info!("Max number of retries ({}) without adding point", MAX_RETRY);
-                    break;
+                    info!("Use LHS optimization to get at least one point");
+                    lhs_optim = true;
                 }
                 info!("End iteration {}/{}", i, n_iter);
-            } else {
-                no_point_added_retries = MAX_RETRY;
-                let count = (self.q_parallel - rejected_count) as i32;
-                let x_to_eval = x_data.slice(s![-count.., ..]).to_owned();
-                info!(
-                    "Add {} point{}:",
-                    count,
-                    if rejected_count > 1 { "s" } else { "" }
-                );
-                info!("  {}", x_dat);
-                let y_actual = self.eval(&x_to_eval);
[…]
