### 2022-06-05 951e3a8
Test and fix auto mode
