### 2022-06-05 08e6a10
remove logging in test
