### 2022-06-08 f6fcee7
Implement no blas using linfa_linag (default)
