+++ gp/src/algorithm.rs
@@ -4,11 +4,15 @@ use crate::mean_models::*;
 use crate::parameters::{GpParams, GpValidParams};
 use crate::utils::{DistanceMatrix, NormalizedMatrix};
 use egobox_doe::{Lhs, SamplingMethod};
+#[cfg(feature = "blas")]
 use linfa::dataset::{WithLapack, WithoutLapack};
 use linfa::prelude::{Dataset, DatasetBase, Fit, Float};
+#[cfg(not(feature = "blas"))]
+use linfa_linalg::{cholesky::*, qr::*, svd::*, triangular::*};
 use linfa_pls::PlsRegression;
 use ndarray::{arr1, s, Array, Array1, Array2, ArrayBase, Axis, Data, Ix2, Zip};
 use ndarray_einsum_beta::*;
+#[cfg(feature = "blas")]
 use ndarray_linalg::{cholesky::*, qr::*, svd::*, triangular::*};
 use ndarray_rand::rand::SeedableRng;
 use ndarray_stats::QuantileExt;
@@ -131,13 +135,18 @@ impl<F: Float, Mean: RegressionModel<F>, Corr: CorrelationModel<F>> GaussianProc
         let inners = &self.inner_params;
 
         let corr_t = corr.t().to_owned();
+        #[cfg(feature = "blas")]
         let rt = inners
             .r_chol
             .to_owned()
             .with_lapack()
             .solve_triangular(UPLO::Lower, Diag::NonUnit, &corr_t.with_lapack())?
             .without_lapack();
+        #[cfg(not(feature = "blas"))]
+        let rt = inners.r_chol.solve_triangular(&corr_t, UPLO::Lower)?;
+
         let lhs = inners.ft.t().dot(&rt) - self.mean.apply(x).t();
+        #[cfg(feature = "blas")]
         let u = inners
             .ft_qr_r
             .to_owned()
@@ -145,6 +154,8 @@ impl<F: Float, Mean: RegressionModel<F>, Corr: CorrelationModel<F>> GaussianProc
             .with_lapack()
             .solve_triangular(UPLO::Upper, Diag::NonUnit, &lhs.with_lapack())?
             .without_lapack();
+        #[cfg(not(feature = "blas"))]
+        let u = inners.ft_qr_r.t().solve_triangular(&lhs, UPLO::Lower)?;
 
         let a = &inners.sigma2;
         let b = Array::ones(rt.ncols()) - rt.mapv(|v| v * v).sum_axis(Axis(0))
@@ -391,16 +402,32 @@ fn reduced_likelihood<F: Float>(
         r_mx[[ij[1], ij[0]]] = rxx[[i, 0]];
     }
 
+    #[cfg(feature = "blas")]
     let fxl = fx.to_owned().with_lapack();
+    #[cfg(not(feature = "blas"))]
+    let fxl = fx;
 
     // R cholesky decomposition
+    #[cfg(feature = "blas")]
     let r_chol = r_mx.with_lapack().cholesky(UPLO::Lower)?;
+    #[cfg(not(feature = "blas"))]
+    let r_chol = r_mx.cholesky()?;
 
     // Solve generalized least squared problem
+    #[cfg(feature = "blas")]
     let ft = r_chol.solve_triangular(UPLO::Lower, Diag::NonUnit, &fxl)?;
+    #[cfg(not(feature = "blas"))]
+    let ft = r_chol.solve_triangular(fxl, UPLO::Lower)?;
+
+    #[cfg(feature = "blas")]
     let (ft_qr_q, ft_qr_r) = ft.qr().unwrap();
+    #[cfg(not(feature = "blas"))]
+    let (ft_qr_q, ft_qr_r) = ft.qr().unwrap().into_decomp();
 
[…]
