### 2022-06-08 e202525
Fix linalg error
