+++ moe/Cargo.toml
@@ -13,6 +13,7 @@ categories = ["algorithms", "mathematics", "science"]
 default = []
 
 persistent = ["serde", "typetag", "egobox-gp/serializable", "serde_json"]
+blas = ["ndarray-linalg", "linfa/ndarray-linalg", "linfa-clustering/blas", "linfa-pls/blas"]
 
 [dependencies]
 egobox-doe = { version = "0.3.0", path = "../doe" }
