### 2022-06-18 85aadbf
Use linfa 0.6.0
