### 2022-06-18 c91b2a2
Test no blas and blas in CI
