### 2022-06-18 9b390ff
Re-enable power method linfa_pls error
