+++ gp/src/algorithm.rs
@@ -230,10 +230,9 @@ impl<F: Float, Mean: RegressionModel<F>, Corr: CorrelationModel<F>, D: Data<Elem
             let ds = Dataset::new(x.to_owned(), y.to_owned());
             w_star = PlsRegression::params(*n_components).fit(&ds).map_or_else(
                 |e| match e {
-                    // FIXME: Commented out in order to compile with linfa 0.5.1, should be ok with next release
-                    // PlsError::PowerMethodConstantResidualError() => {
-                    //     Ok(Array2::zeros((x.ncols(), *n_components)))
-                    // }
+                    linfa_pls::PlsError::PowerMethodConstantResidualError() => {
+                        Ok(Array2::zeros((x.ncols(), *n_components)))
+                    }
                     err => Err(err),
                 },
                 |v| Ok(v.rotations().0.to_owned()),
@@ -532,28 +531,28 @@ mod tests {
     use rand_isaac::Isaac64Rng;
 
     // FIXME: Removed nn order to pass with linfa 0.5.1, should be ok with next release
-    // #[test]
-    // fn test_constant_function() {
-    //     let dim = 3;
-    //     let lim = array![[0., 1.]];
-    //     let xlimits = lim.broadcast((dim, 2)).unwrap();
-    //     let rng = Isaac64Rng::seed_from_u64(42);
-    //     let nt = 30;
-    //     let xt = Lhs::new(&xlimits).with_rng(rng).sample(nt);
-    //     let yt = Array::from_vec(vec![3.1; nt]).insert_axis(Axis(1));
-    //     let gp = GaussianProcess::<f64, ConstantMean, SquaredExponentialCorr>::params(
-    //         ConstantMean::default(),
-    //         SquaredExponentialCorr::default(),
-    //     )
-    //     .initial_theta(Some(vec![0.1]))
-    //     .kpls_dim(Some(1))
-    //     .fit(&Dataset::new(xt, yt))
-    //     .expect("GP fit error");
-    //     let rng = Isaac64Rng::seed_from_u64(43);
-    //     let xtest = Lhs::new(&xlimits).with_rng(rng).sample(nt);
-    //     let ytest = gp.predict_values(&xtest).expect("prediction error");
-    //     assert_abs_diff_eq!(Array::from_elem((nt, 1), 3.1), ytest, epsilon = 1e-6);
-    // }
+    #[test]
+    fn test_constant_function() {
+        let dim = 3;
+        let lim = array![[0., 1.]];
+        let xlimits = lim.broadcast((dim, 2)).unwrap();
+        let rng = Isaac64Rng::seed_from_u64(42);
+        let nt = 30;
+        let xt = Lhs::new(&xlimits).with_rng(rng).sample(nt);
+        let yt = Array::from_vec(vec![3.1; nt]).insert_axis(Axis(1));
+        let gp = GaussianProcess::<f64, ConstantMean, SquaredExponentialCorr>::params(
+            ConstantMean::default(),
+            SquaredExponentialCorr::default(),
+        )
+        .initial_theta(Some(vec![0.1]))
+        .kpls_dim(Some(1))
+        .fit(&Dataset::new(xt, yt))
+        .expect("GP fit error");
+        let rng = Isaac64Rng::seed_from_u64(43);
+        let xtest = Lhs::new(&xlimits).with_rng(rng).sample(nt);
+        let ytest = gp.predict_values(&xtest).expect("prediction error");
+        assert_abs_diff_eq!(Array::from_elem((nt, 1), 3.1), ytest, epsilon = 1e-6);
+    }
 
     macro_rules! test_gp {
         ($regr:ident, $corr:ident, $expected:expr) => {
