+++ ego/src/egor.rs
@@ -921,7 +921,7 @@ mod tests {
         println!("Rosenbrock optim result = {:?}", res);
         println!("Elapsed = {:?}", now.elapsed());
         let expected = array![1., 1.];
-        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 2e-1);
+        assert_abs_diff_eq!(expected, res.x_opt, epsilon = 5e-1);
     }
 
     // Objective
