+++ .github/workflows/test.yml
@@ -48,7 +48,7 @@ jobs:
 
 
   test-blas:
-    name: testing-blas-${{ matrix.toolchain }}-${{ matrix.os }}
+    name: testing-with-blas-${{ matrix.toolchain }}-${{ matrix.os }}
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
@@ -57,6 +57,7 @@ jobs:
           - stable
         os: 
           - ubuntu-18.04
+          - windows-2019
 
     steps:
       - name: Checkout sources
@@ -71,6 +72,13 @@ jobs:
 
       - name: Log active toolchain
         run: rustup show
+
+      - name: Install LLVM and Clang on Windows
+        if: ${{ matrix.os == 'windows-2019' }} 
+        uses: KyleMayes/install-llvm-action@v1
+        with:
+          version: "11.0"
+          directory: ${{ runner.temp }}/llvm
 
       - name: Run cargo test in release mode
         uses: actions-rs/cargo@v1
