### 2022-06-19 871545d
Make BLAS dependency optional (#23)
* Implement no blas using linfa_linag (default)

* Remove useless dependencies

* Fix linalg error

* Add blas feature

* Make clippy happy

* Use linfa 0.6.0

* Test no blas and blas in CI

* Re-enable power method linfa_pls error

* Adjust tests
