+++ gp/src/algorithm.rs
@@ -4,11 +4,15 @@ use crate::mean_models::*;
 use crate::parameters::{GpParams, GpValidParams};
 use crate::utils::{DistanceMatrix, NormalizedMatrix};
 use egobox_doe::{Lhs, SamplingMethod};
+#[cfg(feature = "blas")]
 use linfa::dataset::{WithLapack, WithoutLapack};
 use linfa::prelude::{Dataset, DatasetBase, Fit, Float};
+#[cfg(not(feature = "blas"))]
+use linfa_linalg::{cholesky::*, qr::*, svd::*, triangular::*};
 use linfa_pls::PlsRegression;
 use ndarray::{arr1, s, Array, Array1, Array2, ArrayBase, Axis, Data, Ix2, Zip};
 use ndarray_einsum_beta::*;
+#[cfg(feature = "blas")]
 use ndarray_linalg::{cholesky::*, qr::*, svd::*, triangular::*};
 use ndarray_rand::rand::SeedableRng;
 use ndarray_stats::QuantileExt;
@@ -131,13 +135,18 @@ impl<F: Float, Mean: RegressionModel<F>, Corr: CorrelationModel<F>> GaussianProc
         let inners = &self.inner_params;
 
         let corr_t = corr.t().to_owned();
+        #[cfg(feature = "blas")]
         let rt = inners
             .r_chol
             .to_owned()
             .with_lapack()
             .solve_triangular(UPLO::Lower, Diag::NonUnit, &corr_t.with_lapack())?
             .without_lapack();
+        #[cfg(not(feature = "blas"))]
+        let rt = inners.r_chol.solve_triangular(&corr_t, UPLO::Lower)?;
+
         let lhs = inners.ft.t().dot(&rt) - self.mean.apply(x).t();
+        #[cfg(feature = "blas")]
         let u = inners
             .ft_qr_r
             .to_owned()
@@ -145,6 +154,8 @@ impl<F: Float, Mean: RegressionModel<F>, Corr: CorrelationModel<F>> GaussianProc
             .with_lapack()
             .solve_triangular(UPLO::Upper, Diag::NonUnit, &lhs.with_lapack())?
             .without_lapack();
+        #[cfg(not(feature = "blas"))]
+        let u = inners.ft_qr_r.t().solve_triangular(&lhs, UPLO::Lower)?;
 
         let a = &inners.sigma2;
         let b = Array::ones(rt.ncols()) - rt.mapv(|v| v * v).sum_axis(Axis(0))
@@ -219,10 +230,9 @@ impl<F: Float, Mean: RegressionModel<F>, Corr: CorrelationModel<F>, D: Data<Elem
             let ds = Dataset::new(x.to_owned(), y.to_owned());
             w_star = PlsRegression::params(*n_components).fit(&ds).map_or_else(
                 |e| match e {
-                    // FIXME: Commented out in order to compile with linfa 0.5.1, should be ok with next release
-                    // PlsError::PowerMethodConstantResidualError() => {
-                    //     Ok(Array2::zeros((x.ncols(), *n_components)))
-                    // }
+                    linfa_pls::PlsError::PowerMethodConstantResidualError() => {
+                        Ok(Array2::zeros((x.ncols(), *n_components)))
+                    }
                     err => Err(err),
                 },
                 |v| Ok(v.rotations().0.to_owned()),
@@ -391,16 +401,32 @@ fn reduced_likelihood<F: Float>(
         r_mx[[ij[1], ij[0]]] = rxx[[i, 0]];
     }
 
[…]
