+++ doe/Cargo.toml
@@ -13,7 +13,7 @@ categories = ["algorithms", "mathematics", "science"]
 default = []
 
 [dependencies]
-linfa = { version = "0.5.1", default-features = false }
+linfa = { version = "0.6.0", default-features = false }
 ndarray = { version = "0.15", features = ["rayon", "approx"]}
 ndarray-stats = "0.5"
 ndarray-rand = "0.14"

[…]
