### 2022-06-19 c9d6974
[moe] Automatic number of clusters (#22)
* Make nb clusters optional default to 1

* Fixes

* Test and fix auto mode

* remove logging in test
