+++ moe/examples/clustering.rs
@@ -22,8 +22,11 @@ fn main() -> Result<(), Box<dyn Error>> {
     let xtrain = Lhs::new(&arr2(&[[0., 1.]])).sample(50);
     let ytrain = f3parts(&xtrain);
     let ds = Dataset::new(xtrain, ytrain);
-    let moe1 = Moe::params(1).fit(&ds)?;
-    let moe3 = Moe::params(3).recombination(Recombination::Hard).fit(&ds)?;
+    let moe1 = Moe::params().fit(&ds)?;
+    let moe3 = Moe::params()
+        .n_clusters(3)
+        .recombination(Recombination::Hard)
+        .fit(&ds)?;
 
     let xtest = Array::linspace(0., 1., 101).insert_axis(Axis(1));
     let ytest = f3parts(&xtest);
