+++ moe/src/algorithm.rs
@@ -1,4 +1,5 @@
 use super::gaussian_mixture::GaussianMixture;
+use crate::clustering::find_best_number_of_clusters;
 use crate::errors::MoeError;
 use crate::errors::Result;
 use crate::expertise_macros::*;
@@ -6,13 +7,12 @@ use crate::parameters::{
     CorrelationSpec, MoeParams, MoeValidParams, Recombination, RegressionSpec,
 };
 use crate::surrogates::*;
-use log::{debug, info, trace};
-
 use egobox_gp::{correlation_models::*, mean_models::*, GaussianProcess};
 use linfa::dataset::Records;
 use linfa::traits::{Fit, Predict};
 use linfa::{Dataset, DatasetBase, Float, ParamGuard};
 use linfa_clustering::GaussianMixtureModel;
+use log::{debug, info, trace};
 use paste::paste;
 use std::cmp::Ordering;
 use std::ops::Sub;
@@ -71,8 +71,26 @@ impl<R: Rng + SeedableRng + Clone> MoeValidParams<f64, R> {
         let training;
         let (mut xtest, mut ytest) = (None, None);
 
-        if let Recombination::Smooth(None) = self.recombination() {
-            // Extract 10% of data for validation
+        let (n_clusters, recomb) = if self.n_clusters() == 0 {
+            // automatic mode
+            let max_nb_clusters = xt.nrows() / 10 + 1;
+            find_best_number_of_clusters(
+                xt,
+                yt,
+                max_nb_clusters,
+                self.kpls_dim(),
+                self.regression_spec(),
+                self.correlation_spec(),
+                self.rng(),
+            )
+        } else {
+            (self.n_clusters(), self.recombination())
+        };
+        if self.n_clusters() == 0 {
+            debug!("Automatic settings {} {:?}", n_clusters, recomb);
+        }
+        if let Recombination::Smooth(None) = recomb {
+            // Extract 5% of data for validation
             // TODO: Use cross-validation ? Performances
             let (test, training_data) = extract_part(&data, 5);
             // write_npy("xt_hard.npy", &training.slice(s![.., ..nx])).expect("obs saved");
@@ -86,9 +104,8 @@ impl<R: Rng + SeedableRng + Clone> MoeValidParams<f64, R> {
         }
         let dataset = Dataset::from(training);
 
-        let gmm = GaussianMixtureModel::params(self.n_clusters())
+        let gmm = GaussianMixtureModel::params(n_clusters)
             .n_runs(20)
-            //.reg_covariance(1e-6)
             .with_rng(self.rng())
             .fit(&dataset)
             .expect("Training data clustering");
@@ -97,7 +114,7 @@ impl<R: Rng + SeedableRng + Clone> MoeValidParams<f64, R> {
         let weights = gmm.weights().to_owned();
         let means = gmm.means().slice(s![.., ..nx]).to_owned();
         let covariances = gmm.covariances().slice(s![.., ..nx, ..nx]).to_owned();
-        let factor = match self.recombination() {
+        let factor = match recomb {
             Recombination::Smooth(Some(f)) => f,
             Recombination::Smooth(None) => 1.,
             Recombination::Hard => 1.,
@@ -105,7 +122,7 @@ impl<R: Rng + SeedableRng + Clone> MoeValidParams<f64, R> {
         let gmx = GaussianMixture::new(weights, means, covariances)?.with_heaviside_factor(factor);
 
         let dataset_clustering = gmx.predict(xt);
-        let clusters = sort_by_cluster(self.n_clusters(), &data, &dataset_clustering);
[…]
