### 2022-06-19 7fecd95
Make LhsOptimizer reproducible (rng setting)
