### 2022-06-19 9574f74
Make LhsOptimizer reproducible (rng setting)
