+++ src/lib.rs
@@ -257,8 +257,9 @@ impl Vspec {
 ///         Number of components to be used specifiying PLS projection is used (a.k.a KPLS method).
 ///         This is used to address high-dimensional problems typically when nx > 9.
 ///
-///     n_clusters (int > 0)
+///     n_clusters (int >= 0)
 ///         Number of clusters used by the mixture of surrogate experts.
+///         When set to 0, the number of cluster is determined automatically.
 ///   
 ///     expected (ExpectedOptimum)
 ///         Known optimum used as stopping criterion.
