+++ ego/src/egor.rs
@@ -99,7 +99,7 @@ use crate::utils::{compute_cstr_scales, compute_obj_scale, compute_wb2s_scale};
 use crate::utils::{ei, wb2s};
 
 use egobox_doe::{Lhs, LhsKind, SamplingMethod};
-use egobox_moe::{CorrelationSpec, Moe, MoeParams, RegressionSpec, Surrogate};
+use egobox_moe::{ClusteredSurrogate, CorrelationSpec, Moe, MoeParams, RegressionSpec};
 use env_logger::{Builder, Env};
 use finitediff::FiniteDiff;
 use linfa::{traits::Fit, Dataset, ParamGuard};
@@ -115,10 +115,10 @@ const DOE_INITIAL_FILE: &str = "egor_initial_doe.npy";
 const DOE_FILE: &str = "egobox_doe.npy";
 
 impl<R: Rng + SeedableRng + Clone> SurrogateBuilder for MoeParams<f64, R> {
-    fn train(&self, xt: &Array2<f64>, yt: &Array2<f64>) -> Result<Box<dyn Surrogate>> {
+    fn train(&self, xt: &Array2<f64>, yt: &Array2<f64>) -> Result<Box<dyn ClusteredSurrogate>> {
         let checked = self.check_ref()?;
         let moe = checked.fit(&Dataset::new(xt.to_owned(), yt.to_owned()))?;
-        Ok(moe).map(|moe| Box::new(moe) as Box<dyn Surrogate>)
+        Ok(moe).map(|moe| Box::new(moe) as Box<dyn ClusteredSurrogate>)
     }
 }
 
@@ -394,7 +394,8 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
     pub fn suggest(&self, x_data: &Array2<f64>, y_data: &Array2<f64>) -> Array2<f64> {
         let rng = self.rng.clone();
         let sampling = Lhs::new(&self.xlimits).with_rng(rng).kind(LhsKind::Maximin);
-        let (x_dat, _) = self.next_points(0, x_data, y_data, &sampling, false);
+        let mut n_clusters = vec![self.n_clusters.unwrap_or(1); 1 + self.n_cstr];
+        let (x_dat, _) = self.next_points(0, &mut n_clusters, x_data, y_data, &sampling, false);
         x_dat
     }
 
@@ -460,10 +461,13 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
         }
 
         const MAX_RETRY: i32 = 3;
+        let mut n_clusters = vec![self.n_clusters.unwrap_or(1); 1 + self.n_cstr];
         let mut no_point_added_retries = MAX_RETRY;
         let mut lhs_optim = false;
+
         for i in 1..=n_iter {
-            let (x_dat, y_dat) = self.next_points(i, &x_data, &y_data, &sampling, lhs_optim);
+            let (x_dat, y_dat) =
+                self.next_points(i, &mut n_clusters, &x_data, &y_data, &sampling, lhs_optim);
             debug!("Try adding {}", x_dat);
             let rejected_count = update_data(&mut x_data, &mut y_data, &x_dat, &y_dat);
             if rejected_count > 0 {
@@ -536,9 +540,37 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
         Ok(res)
     }
 
+    fn should_recluster(&self, n_iter: usize) -> bool {
+        if let Some(nc) = self.n_clusters {
+            n_iter % 10 == 1 && nc == 0
+        } else {
+            false
+        }
+    }
+
+    fn make_default_builder(
+        &self,
+        clustering_i: usize,
[…]
