+++ moe/src/lib.rs
@@ -106,6 +106,7 @@ mod parameters;
 mod surrogates;
 
 pub use algorithm::*;
+pub use clustering::*;
 pub use errors::*;
 pub use parameters::*;
 pub use surrogates::*;
