+++ ego/src/utils.rs
@@ -1,4 +1,4 @@
-use egobox_moe::Surrogate;
+use egobox_moe::ClusteredSurrogate;
 use libm::erfc;
 use ndarray::{
     concatenate, Array1, Array2, ArrayBase, ArrayView, ArrayView2, Axis, Data, Ix1, Ix2, Zip,
@@ -10,7 +10,7 @@ use ndarray_stats::{DeviationExt, QuantileExt};
 
 const SQRT_2PI: f64 = 2.5066282746310007;
 
-pub fn ei(x: &[f64], obj_model: &dyn Surrogate, f_min: f64) -> f64 {
+pub fn ei(x: &[f64], obj_model: &dyn ClusteredSurrogate, f_min: f64) -> f64 {
     let pt = ArrayView::from_shape((1, x.len()), x).unwrap();
     if let Ok(p) = obj_model.predict_values(&pt) {
         if let Ok(s) = obj_model.predict_variances(&pt) {
@@ -28,13 +28,17 @@ pub fn ei(x: &[f64], obj_model: &dyn Surrogate, f_min: f64) -> f64 {
     }
 }
 
-pub fn wb2s(x: &[f64], obj_model: &dyn Surrogate, f_min: f64, scale: f64) -> f64 {
+pub fn wb2s(x: &[f64], obj_model: &dyn ClusteredSurrogate, f_min: f64, scale: f64) -> f64 {
     let pt = ArrayView::from_shape((1, x.len()), x).unwrap();
     let ei = ei(x, obj_model, f_min);
     scale * ei - obj_model.predict_values(&pt).unwrap()[[0, 0]]
 }
 
-pub fn compute_wb2s_scale(x: &ArrayView2<f64>, obj_model: &dyn Surrogate, f_min: f64) -> f64 {
+pub fn compute_wb2s_scale(
+    x: &ArrayView2<f64>,
+    obj_model: &dyn ClusteredSurrogate,
+    f_min: f64,
+) -> f64 {
     let ratio = 100.; // TODO: make it a parameter
     let ei_x = x.map_axis(Axis(1), |xi| {
         let ei = ei(xi.as_slice().unwrap(), obj_model, f_min);
@@ -52,12 +56,15 @@ pub fn compute_wb2s_scale(x: &ArrayView2<f64>, obj_model: &dyn Surrogate, f_min:
     }
 }
 
-pub fn compute_obj_scale(x: &ArrayView2<f64>, obj_model: &dyn Surrogate) -> f64 {
+pub fn compute_obj_scale(x: &ArrayView2<f64>, obj_model: &dyn ClusteredSurrogate) -> f64 {
     let preds = obj_model.predict_values(x).unwrap().mapv(f64::abs);
     *preds.max().unwrap_or(&1.0)
 }
 
-pub fn compute_cstr_scales(x: &ArrayView2<f64>, cstr_models: &[Box<dyn Surrogate>]) -> Array1<f64> {
+pub fn compute_cstr_scales(
+    x: &ArrayView2<f64>,
+    cstr_models: &[Box<dyn ClusteredSurrogate>],
+) -> Array1<f64> {
     let scales: Vec<f64> = cstr_models
         .iter()
         .map(|cstr_model| {
