### 2022-06-23 9745073
Fix gp updates in qEI implementation and adjust eval nb
