### 2022-06-23 1ec099a
Fix qEI strategy handling (#26)
* Fix docstrings

* Fix gp updates in qEI implementation and adjust eval nb

* Adjust nb of iterations display
