### 2022-06-24 8568b0f
Fix cstr clustering indexing
