### 2022-06-24 f221043
Use specified nb clusters
