### 2022-06-24 53b3df9
Avoid clustering when factor is optimized
