### 2022-06-25 d5a0cb3
Implement automatic reclustering (#25)
* Implement automatic reclustering

* Compilation ok

* Fix cstr clustering indexing

* Fix moe training

* Fix doc test

* Use specified nb clusters

* Avoid clustering when factor is optimized

* Revert previous change: needs to retrain on whole data
