+++ ego/src/mixint.rs
@@ -5,8 +5,10 @@
 use crate::errors::{EgoError, Result};
 use crate::types::SurrogateBuilder;
 use egobox_doe::{Lhs, SamplingMethod};
-use egobox_moe::{Moe, MoeParams, RegressionSpec, Surrogate};
-use linfa::{traits::Fit, Dataset, DatasetBase};
+use egobox_moe::{
+    Clustered, ClusteredSurrogate, Clustering, Moe, MoeParams, RegressionSpec, Surrogate,
+};
+use linfa::{traits::Fit, DatasetBase, ParamGuard};
 use ndarray::{s, Array, Array2, ArrayView2, Axis, Zip};
 use ndarray_rand::rand::SeedableRng;
 use ndarray_stats::QuantileExt;
@@ -229,9 +231,8 @@ impl SamplingMethod<f64> for MixintSampling {
 
 /// Moe type for MixintEgor optimizer
 pub type MoeBuilder = MoeParams<f64, Isaac64Rng>;
-
 /// A decorator of Moe surrogate that takes into account Xtype specifications
-pub struct MixintMoeParams {
+pub struct MixintMoeValidParams {
     /// The surrogate factory
     surrogate_builder: MoeBuilder,
     /// The input specifications
@@ -241,20 +242,10 @@ pub struct MixintMoeParams {
     work_in_folded_space: bool,
 }
 
-impl MixintMoeParams {
-    /// Constructor given  `xtypes` specifications and given surrogate builder
-    pub fn new(xtypes: &[Xtype], surrogate_builder: &MoeBuilder) -> Self {
-        MixintMoeParams {
-            surrogate_builder: surrogate_builder.clone(),
-            xtypes: xtypes.to_vec(),
-            work_in_folded_space: false,
-        }
-    }
-
+impl MixintMoeValidParams {
     /// Sets whether we want to work in folded space
-    pub fn work_in_folded_space(&mut self, wfs: bool) -> &mut Self {
-        self.work_in_folded_space = wfs;
-        self
+    pub fn work_in_folded_space(&self) -> bool {
+        self.work_in_folded_space
     }
 
     /// Sets the specification
@@ -263,35 +254,122 @@ impl MixintMoeParams {
     }
 }
 
-impl SurrogateBuilder for MixintMoeParams {
-    fn train(&self, x: &Array2<f64>, y: &Array2<f64>) -> Result<Box<dyn Surrogate>> {
-        Ok(Box::new(self.fit(&Dataset::new(x.to_owned(), y.to_owned()))?) as Box<dyn Surrogate>)
+pub struct MixintMoeParams(MixintMoeValidParams);
+
+impl MixintMoeParams {
+    /// Constructor given  `xtypes` specifications and given surrogate builder
+    pub fn new(xtypes: &[Xtype], surrogate_builder: &MoeBuilder) -> Self {
+        MixintMoeParams(MixintMoeValidParams {
+            surrogate_builder: surrogate_builder.clone(),
+            xtypes: xtypes.to_vec(),
+            work_in_folded_space: false,
+        })
+    }
+
+    /// Sets whether we want to work in folded space
+    pub fn work_in_folded_space(&mut self, wfs: bool) -> &mut Self {
+        self.0.work_in_folded_space = wfs;
+        self
+    }
+
+    /// Sets the specification
+    pub fn xtypes(&self) -> &[Xtype] {
+        &self.0.xtypes
     }
 }
 
-impl Fit<Array2<f64>, Array2<f64>, EgoError> for MixintMoeParams {
+impl MixintMoeValidParams {
+    fn _train(&self, xt: &Array2<f64>, yt: &Array2<f64>) -> Result<MixintMoe> {
[…]
