+++ moe/src/lib.rs
@@ -106,6 +106,8 @@ mod parameters;
 mod surrogates;
 
 pub use algorithm::*;
+pub use clustering::*;
 pub use errors::*;
+pub use gaussian_mixture::*;
 pub use parameters::*;
 pub use surrogates::*;
