+++ src/lib.rs
@@ -260,7 +260,9 @@ impl Vspec {
 ///
 ///     n_clusters (int >= 0)
 ///         Number of clusters used by the mixture of surrogate experts.
-///         When set to 0, the number of cluster is determined automatically.
+///         When set to 0, the number of cluster is determined automatically and refreshed every
+///         10-points addition (should say 'tentative addition' because addition may fail for some points
+///         but it is counted anyway).
 ///   
 ///     expected (ExpectedOptimum)
 ///         Known optimum used as stopping criterion.
