### 2022-06-29 4d3a77a
Improve nb of eval management and logging, fix LHS optim seeding
