+++ ego/src/egor.rs
@@ -135,7 +135,7 @@ impl<R: Rng + SeedableRng + Clone> SurrogateBuilder for MoeParams<f64, R> {
 
 /// EGO optimization parameterization
 #[derive(Clone)]
-pub struct Egor<'a, O: GroupFunc, R: Rng> {
+pub struct Egor<'a, O: GroupFunc, R: SeedableRng> {
     /// Number of function evaluations allocated to find the optimum
     /// Note: if the initial doe has to be evaluated, doe size should be substract.  
     pub n_eval: usize,
@@ -204,7 +204,7 @@ impl<'a, O: GroupFunc> Egor<'a, O, Isaac64Rng> {
     }
 }
 
-impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
+impl<'a, O: GroupFunc, R: Rng + SeedableRng + Clone> Egor<'a, O, R> {
     /// Constructor of the optimization of the function `f` with specified random generator
     /// to get reproducibility.
     ///
@@ -375,7 +375,7 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
     }
 
     /// Sets a random generator for reproducibility
-    pub fn with_rng<R2: Rng + Clone>(self, rng: R2) -> Egor<'a, O, R2> {
+    pub fn with_rng<R2: Rng + SeedableRng + Clone>(self, rng: R2) -> Egor<'a, O, R2> {
         Egor {
             n_eval: self.n_eval,
             n_start: self.n_start,
@@ -410,7 +410,15 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
         let rng = self.rng.clone();
         let sampling = Lhs::new(&self.xlimits).with_rng(rng).kind(LhsKind::Maximin);
         let mut clusterings = vec![None; 1 + self.n_cstr];
-        let (x_dat, _) = self.next_points(1, &mut clusterings, x_data, y_data, &sampling, false);
+        let (x_dat, _) = self.next_points(
+            true,
+            false, // done anyway
+            &mut clusterings,
+            x_data,
+            y_data,
+            &sampling,
+            None,
+        );
         x_dat
     }
 
@@ -438,7 +446,7 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
 
         let doe = hstart_doe.as_ref().or(self.doe.as_ref());
 
-        let (mut y_data, mut x_data, n_iter) = if let Some(doe) = doe {
+        let (mut y_data, mut x_data, n_eval) = if let Some(doe) = doe {
             if doe.ncols() == self.xlimits.nrows() {
                 // only x are specified
                 info!("Compute initial DOE on specified {} points", doe.nrows());
@@ -453,7 +461,7 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
                 (
                     doe.slice(s![.., self.xlimits.nrows()..]).to_owned(),
                     doe.slice(s![.., ..self.xlimits.nrows()]).to_owned(),
-                    self.n_eval.saturating_sub(doe.nrows()),
+                    self.n_eval,
                 )
             }
         } else {
@@ -476,21 +484,66 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
         }
 
         const MAX_RETRY: i32 = 3;
-        info!("NUMBER CLUSTER {:?}", self.n_clusters);
         let mut clusterings = vec![None; self.n_cstr + 1];
         let mut no_point_added_retries = MAX_RETRY;
-        let mut lhs_optim = false;
-        if n_iter / self.q_parallel == 0 {
[…]
