### 2022-07-09 3e4b1e4
Avoid infinite loop if point addition fails
