+++ .github/workflows/pytest.yml
@@ -51,7 +51,7 @@ jobs:
       - name: Test Python
         run: |
             poetry install
-            poetry run maturin develop --release --cargo-extra-args="--features linfa/intel-mkl-static"
+            poetry run maturin develop --release
             poetry run pytest
         shell: bash
 
