### 2022-07-09 74ed32f
Fix reject points continue logic
