### 2022-07-09 4ff9bda
Release 0.4.0 (#27)
* Make LhsOptimizer reproducible (rng setting)

* Update 0.4.0 general info

* Document nb cluster sets to 0

* Document reclustering

* Improve reclustering logging

* Improve nb of eval management and logging, fix LHS optim seeding

* Shut up clippy

* Cleanup test files

* Avoid infinite loop if point addition fails

* Build python package without blas

* Fix reject points continue logic

* Increase iteration to make test pass

* Bump 0.4.0
