+++ ego/src/egor.rs
@@ -135,7 +135,7 @@ impl<R: Rng + SeedableRng + Clone> SurrogateBuilder for MoeParams<f64, R> {
 
 /// EGO optimization parameterization
 #[derive(Clone)]
-pub struct Egor<'a, O: GroupFunc, R: Rng> {
+pub struct Egor<'a, O: GroupFunc, R: SeedableRng> {
     /// Number of function evaluations allocated to find the optimum
     /// Note: if the initial doe has to be evaluated, doe size should be substract.  
     pub n_eval: usize,
@@ -169,6 +169,8 @@ pub struct Egor<'a, O: GroupFunc, R: Rng> {
     /// Optional dimension reduction (see [egobox_moe])
     pub kpls_dim: Option<usize>,
     /// Number of clusters used by mixture of experts (see [egobox_moe])
+    /// When set to 0 the clusters are computes automatically and refreshed
+    /// every 10-points (tentative) additions
     pub n_clusters: Option<usize>,
     /// Specification of an expected solution which is used to stop the algorithm once reached
     pub expected: Option<ApproxValue>,
@@ -202,7 +204,7 @@ impl<'a, O: GroupFunc> Egor<'a, O, Isaac64Rng> {
     }
 }
 
-impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
+impl<'a, O: GroupFunc, R: Rng + SeedableRng + Clone> Egor<'a, O, R> {
     /// Constructor of the optimization of the function `f` with specified random generator
     /// to get reproducibility.
     ///
@@ -328,6 +330,8 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
     }
 
     /// Sets the number of clusters used by the mixture of surrogate experts.
+    ///
+    /// When set to 0, the number of clusters is determined automatically
     pub fn n_clusters(&mut self, n_clusters: Option<usize>) -> &mut Self {
         self.n_clusters = n_clusters;
         self
@@ -371,7 +375,7 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
     }
 
     /// Sets a random generator for reproducibility
-    pub fn with_rng<R2: Rng + Clone>(self, rng: R2) -> Egor<'a, O, R2> {
+    pub fn with_rng<R2: Rng + SeedableRng + Clone>(self, rng: R2) -> Egor<'a, O, R2> {
         Egor {
             n_eval: self.n_eval,
             n_start: self.n_start,
@@ -406,7 +410,15 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
         let rng = self.rng.clone();
         let sampling = Lhs::new(&self.xlimits).with_rng(rng).kind(LhsKind::Maximin);
         let mut clusterings = vec![None; 1 + self.n_cstr];
-        let (x_dat, _) = self.next_points(1, &mut clusterings, x_data, y_data, &sampling, false);
+        let (x_dat, _) = self.next_points(
+            true,
+            false, // done anyway
+            &mut clusterings,
+            x_data,
+            y_data,
+            &sampling,
+            None,
+        );
         x_dat
     }
 
@@ -434,7 +446,7 @@ impl<'a, O: GroupFunc, R: Rng + Clone> Egor<'a, O, R> {
 
         let doe = hstart_doe.as_ref().or(self.doe.as_ref());
 
-        let (mut y_data, mut x_data, n_iter) = if let Some(doe) = doe {
+        let (mut y_data, mut x_data, n_eval) = if let Some(doe) = doe {
             if doe.ncols() == self.xlimits.nrows() {
                 // only x are specified
[…]
