### 2022-07-09 422f5bc
Update CHANGELOG for 0.4.0
