+++ pyproject.toml
@@ -17,8 +17,9 @@ classifiers = [
 build-backend = "maturin"
 requires = ["maturin>=0.12,<0.13", "poetry_core>=1.0.0"]
 
-[tool.maturin]
-cargo-extra-args = "--features linfa/intel-mkl-static"
+# Optional usage of BLAS backend 
+# [tool.maturin]
+# cargo-extra-args = "--features linfa/intel-mkl-static"
 
 [tool.poetry]
 name = "egobox"
