### 2022-07-12 1f68860
Add benchmarks for find_best_nb_clusters
