### 2022-07-12 ba4a868
Remove unused aic/bic computations
