+++ moe/src/clustering.rs
@@ -69,19 +69,15 @@ pub fn find_best_number_of_clusters<R: Rng + Clone>(
     } else {
         max_nb_clusters
     };
-    //let max_nb_clusters = 3;
-    //let val = concatenate(Axis(1), &[x.view(), y.view()]).unwrap();
     let nx = x.ncols();
-
     let dataset: DatasetView<f64, f64> = DatasetView::new(x.view(), y.view());
 
     // Stock
-    let mut errorih: Vec<f64> = Vec::new();
-    let mut erroris: Vec<f64> = Vec::new();
-    let mut posi: Vec<usize> = Vec::new();
-
-    let mut median_eh: Vec<f64> = Vec::new();
-    let mut median_es: Vec<f64> = Vec::new();
+    let mut mean_err_h: Vec<f64> = Vec::new();
+    let mut mean_err_s: Vec<f64> = Vec::new();
+    let mut median_err_h: Vec<f64> = Vec::new();
+    let mut median_err_s: Vec<f64> = Vec::new();
+    let mut nb_clusters_ok: Vec<usize> = Vec::new();
 
     // Init Output Loop
     let mut auxkh;
@@ -163,18 +159,13 @@ pub fn find_best_number_of_clusters<R: Rng + Clone>(
             }
         }
 
-        // Stock median
-        median_es.push(median(&s_errors));
-        median_eh.push(median(&h_errors));
-
-        // Stock possible numbers of cluster
-        if ok {
-            posi.push(i);
-        }
+        // Stock median errors
+        median_err_s.push(median(&s_errors));
+        median_err_h.push(median(&h_errors));
 
         // Stock mean errors
-        erroris.push(mean(&s_errors));
-        errorih.push(mean(&h_errors));
+        mean_err_s.push(mean(&s_errors));
+        mean_err_h.push(mean(&h_errors));
 
         debug!(
             "Number of cluster: {} \
@@ -185,12 +176,17 @@ pub fn find_best_number_of_clusters<R: Rng + Clone>(
             | Median (smooth): {}",
             i + 1,
             ok,
-            mean(&h_errors),
-            mean(&s_errors),
-            median(&h_errors),
-            median(&s_errors),
+            mean_err_h[mean_err_h.len() - 1],
+            mean_err_s[mean_err_s.len() - 1],
+            median_err_h[median_err_h.len() - 1],
+            median_err_s[median_err_s.len() - 1],
         );
         debug!("#######");
+
+        // Stock possible numbers of cluster
+        if ok {
+            nb_clusters_ok.push(i);
+        }
 
         if i > 3 {
             // Stop the search if the clustering can not be performed three times
@@ -201,24 +197,24 @@ pub fn find_best_number_of_clusters<R: Rng + Clone>(
         if use_median {
             // Stop the search if the median increases three times
             if i > 3 {
-                auxkh = median_eh[i - 2];
-                auxkph = median_eh[i - 1];
-                auxkpph = median_eh[i];
-                auxks = median_es[i - 2];
-                auxkps = median_es[i - 1];
-                auxkpps = median_es[i];
+                auxkh = median_err_h[i - 2];
+                auxkph = median_err_h[i - 1];
+                auxkpph = median_err_h[i];
+                auxks = median_err_s[i - 2];
+                auxkps = median_err_s[i - 1];
+                auxkpps = median_err_s[i];
 
                 exit_ =
                     auxkph >= auxkh && auxkps >= auxks && auxkpph >= auxkph && auxkpps >= auxkps;
             }
         } else if i > 3 {
             // Stop the search if the means of errors increase three times
-            auxkh = errorih[i - 2];
[…]
