+++ Cargo.toml
@@ -47,7 +47,6 @@ pyo3 = { version = "0.16.5", features = ["extension-module"] }
 pyo3-log = "0.6.0"
 
 [dev-dependencies]
-linfa-datasets = { version = "0.5.1", features = ["linnerud"] }
 criterion = "0.3"
 serde_json = "1"
 approx = "0.4"
