+++ Cargo.toml
@@ -34,7 +34,7 @@ linfa = { version = "0.6.0", default-features = false }
 
 ndarray = { version = "0.15", features = ["rayon", "approx"]}
 ndarray-rand = "0.14"
-numpy = "0.15.0"
+numpy = "0.16.2"
 anyhow = "=1.0.48"
 
 # nlopt = "0.5.3"
@@ -43,8 +43,8 @@ libm = "0.2"
 finitediff = { version="0.1", features = ["ndarray"] }
 env_logger = "0.9"
 log = "0.4"
-pyo3 = { version = "0.15.1", features = ["extension-module"] }
-pyo3-log = "0.5.0"
+pyo3 = { version = "0.16.5", features = ["extension-module"] }
+pyo3-log = "0.6.0"
 
 [dev-dependencies]
 linfa-datasets = { version = "0.5.1", features = ["linnerud"] }
