### 2022-07-18 e40460b
Use pyo3 native enum capability
