### 2022-07-18 fb558d9
Upgrade pyo3 (#28)
* Upgrade pyo3 0.16

* Use pyo3 native enum capability

* Bump to ubuntu 20.04 (outdated py3.6 in 18..04)
