+++ .github/workflows/test.yml
@@ -16,7 +16,7 @@ jobs:
         toolchain:
           - stable
         os: 
-          - ubuntu-18.04
+          - ubuntu-20.04
           - windows-2019
 
     steps:
@@ -56,7 +56,7 @@ jobs:
         toolchain:
           - stable
         os: 
-          - ubuntu-18.04
+          - ubuntu-20.04
           - windows-2019
 
     steps:
