### 2022-07-18 7ba70f3
Improve clustering (#29)
* Add benchmarks for find_best_nb_clusters

* Remove unused aic/bic computations

* Rename variables

* Remove linfa-datasets dependency
