+++ ego/Cargo.toml
@@ -45,7 +45,6 @@ serde_json = { version = "1", optional = true }
 typetag = { version = "0.1", optional = true }
 
 [dev-dependencies]
-linfa-datasets = { version = "0.5.1", features = ["linnerud"] }
 criterion = "0.3"
 serde_json = "1"
 approx = "0.4"
