### 2022-07-19 0c295c4
Improve logging messages
