### 2022-07-19 91448b6
Implement interruption on Ctrl+C
