### 2022-07-19 a6cdc4a
Manage set_handler error under pytest
