### 2022-07-20 17a4875
Egor interrupt (#30)
* Improve logging messages

* Implement interruption on Ctrl+C

* Manage set_handler error under pytest
