### 2022-08-18 2c1239d
Fix clippy errors
