### 2022-08-21 cafea2e
Put egor wrapping code in a separate module
