+++ src/lib.rs
@@ -1,499 +1,9 @@
 #![doc = include_str!("../README.md")]
 
-use egobox_doe::SamplingMethod;
-use linfa::ParamGuard;
-use log::info;
-use ndarray::{Array2, ArrayView2};
-use ndarray_rand::rand::SeedableRng;
-use numpy::{IntoPyArray, PyArray2, PyReadonlyArray2};
-use pyo3::exceptions::{PyRuntimeError, PyValueError};
+mod egor;
+
+use egor::*;
 use pyo3::prelude::*;
-use rand_isaac::Isaac64Rng;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
-
-/// Utility function converting `xlimits` float data list specifying bounds of x components
-/// to x specified as a list of Vtype.Float types [egobox.Vtype]
-///
-/// # Parameters
-///     xlimits : nx-size list of [lower_bound, upper_bound] where `nx` is the dimension of x
-///
-/// # Returns
-///     xtypes: nx-size list of Vspec(Vtype(FLOAT), [lower_bound, upper_bounds]) where `nx` is the dimension of x
-#[pyfunction]
-fn to_specs(py: Python, xlimits: Vec<Vec<f64>>) -> PyResult<PyObject> {
-    if xlimits.is_empty() || xlimits[0].is_empty() {
-        let err = "Error: xspecs argument cannot be empty";
-        return Err(PyValueError::new_err(err.to_string()));
-    }
-    Ok(xlimits
-        .iter()
-        .map(|xlimit| Vspec::new(Vtype(Vtype::FLOAT), xlimit.clone()))
-        .collect::<Vec<Vspec>>()
-        .into_py(py))
-}
-
-#[pyfunction]
-fn lhs(py: Python, xspecs: PyObject, n_samples: usize, seed: Option<u64>) -> &PyArray2<f64> {
-    let specs: Vec<Vspec> = xspecs.extract(py).expect("Error in xspecs conversion");
-    if specs.is_empty() {
-        panic!("Error: xspecs argument cannot be empty")
-    }
-    let xtypes: Vec<egobox_ego::Xtype> = specs
-        .iter()
-        .map(|spec| match spec.vtype {
-            Vtype(Vtype::FLOAT) => egobox_ego::Xtype::Cont(spec.vlimits[0], spec.vlimits[1]),
-            Vtype(Vtype::INT) => {
-                egobox_ego::Xtype::Int(spec.vlimits[0] as i32, spec.vlimits[1] as i32)
-            }
-            Vtype(i) => panic!(
-                "Bad variable type: should be either Vtype.FLOAT {} or Vtype.INT {}, got {}",
-                Vtype::FLOAT,
-                Vtype::INT,
-                i
-            ),
-        })
-        .collect();
-    let lhs = egobox_ego::MixintContext::new(&xtypes).create_sampling(seed);
-    let doe = lhs.sample(n_samples);
-    doe.into_pyarray(py)
-}
-
-#[pyclass]
-struct RegressionSpec(u8);
-
-#[pymethods]
-impl RegressionSpec {
-    #[classattr]
-    const ALL: u8 = egobox_moe::RegressionSpec::ALL.bits();
-    #[classattr]
-    const CONSTANT: u8 = egobox_moe::RegressionSpec::CONSTANT.bits();
-    #[classattr]
-    const LINEAR: u8 = egobox_moe::RegressionSpec::LINEAR.bits();
-    #[classattr]
-    const QUADRATIC: u8 = egobox_moe::RegressionSpec::QUADRATIC.bits();
-}
-
-#[pyclass]
[…]
