+++ src/egor.rs
@@ -1,0 +1,511 @@
+//! `egobox`, Rust toolbox for efficient global optimization
+//!
+//! Thanks to the [PyO3 project](https://pyo3.rs), which makes Rust well suited for building Python extensions,
+//! the EGO algorithm written in Rust (aka `egor`) is binded in Python. You can install the Python package using:
+//!
+//! ```bash
+//! pip install egobox
+//! ```
+//!
+//! See the [tutorial notebook](https://github.com/relf/egobox/doc/TutorialEgor.ipynb) for usage.
+//!
+
+use egobox_doe::SamplingMethod;
+use linfa::ParamGuard;
+use log::info;
+use ndarray::{Array2, ArrayView2};
+use ndarray_rand::rand::SeedableRng;
+use numpy::{IntoPyArray, PyArray2, PyReadonlyArray2};
+use pyo3::exceptions::{PyRuntimeError, PyValueError};
+use pyo3::prelude::*;
+use rand_isaac::Isaac64Rng;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::Arc;
+
+/// Utility function converting `xlimits` float data list specifying bounds of x components
+/// to x specified as a list of Vtype.Float types [egobox.Vtype]
+///
+/// # Parameters
+///     xlimits : nx-size list of [lower_bound, upper_bound] where `nx` is the dimension of x
+///
+/// # Returns
+///     xtypes: nx-size list of Vspec(Vtype(FLOAT), [lower_bound, upper_bounds]) where `nx` is the dimension of x
+#[pyfunction]
+pub(crate) fn to_specs(py: Python, xlimits: Vec<Vec<f64>>) -> PyResult<PyObject> {
+    if xlimits.is_empty() || xlimits[0].is_empty() {
+        let err = "Error: xspecs argument cannot be empty";
+        return Err(PyValueError::new_err(err.to_string()));
+    }
+    Ok(xlimits
+        .iter()
+        .map(|xlimit| Vspec::new(Vtype(Vtype::FLOAT), xlimit.clone()))
+        .collect::<Vec<Vspec>>()
+        .into_py(py))
+}
+
+#[pyfunction]
+pub(crate) fn lhs(
+    py: Python,
+    xspecs: PyObject,
+    n_samples: usize,
+    seed: Option<u64>,
+) -> &PyArray2<f64> {
+    let specs: Vec<Vspec> = xspecs.extract(py).expect("Error in xspecs conversion");
+    if specs.is_empty() {
+        panic!("Error: xspecs argument cannot be empty")
+    }
+    let xtypes: Vec<egobox_ego::Xtype> = specs
+        .iter()
+        .map(|spec| match spec.vtype {
+            Vtype(Vtype::FLOAT) => egobox_ego::Xtype::Cont(spec.vlimits[0], spec.vlimits[1]),
+            Vtype(Vtype::INT) => {
+                egobox_ego::Xtype::Int(spec.vlimits[0] as i32, spec.vlimits[1] as i32)
+            }
+            Vtype(i) => panic!(
+                "Bad variable type: should be either Vtype.FLOAT {} or Vtype.INT {}, got {}",
+                Vtype::FLOAT,
+                Vtype::INT,
+                i
+            ),
+        })
+        .collect();
+    let lhs = egobox_ego::MixintContext::new(&xtypes).create_sampling(seed);
+    let doe = lhs.sample(n_samples);
+    doe.into_pyarray(py)
+}
+
+#[pyclass]
+pub(crate) struct RegressionSpec(u8);
+
+#[pymethods]
+impl RegressionSpec {
+    #[classattr]
[…]
