### 2022-08-24 2f26c67
Wrap training and prediction
