+++ src/gpmix.rs
@@ -1,22 +1,22 @@
 use crate::types::*;
-use egobox_moe::MoeValidParams;
-use ndarray::Array2;
-use numpy::{PyArray2, PyReadonlyArray2};
+use egobox_moe::Moe;
+use linfa::{traits::Fit, Dataset};
+use ndarray_rand::rand::SeedableRng;
+use numpy::{IntoPyArray, PyArray2, PyReadonlyArray2};
 use pyo3::prelude::*;
 use rand_isaac::Isaac64Rng;
-
 #[pyclass]
 #[pyo3(text_signature = "()")]
 pub(crate) struct GpMix {
-    pub n_clusters: Option<usize>,
+    pub n_clusters: usize,
     pub regression_spec: RegressionSpec,
     pub correlation_spec: CorrelationSpec,
+    pub recombination: Recombination,
     pub kpls_dim: Option<usize>,
-    pub outdir: Option<String>,
+    pub _outdir: Option<String>,
     pub seed: Option<u64>,
-    pub xt: Option<Array2<f64>>,
-    pub yt: Option<Array2<f64>>,
-    pub moe_params: Option<MoeValidParams<f64, Isaac64Rng>>,
+    pub training_data: Option<Dataset<f64, f64>>,
+    pub moe: Option<Moe>,
 }
 
 #[pymethods]
@@ -26,44 +26,87 @@ impl GpMix {
         n_clusters = "1",
         regr_spec = "RegressionSpec::ALL",
         corr_spec = "CorrelationSpec::ALL",
+        recombination = "Recombination::Smooth",
         kpls_dim = "None",
         outdir = "None",
         seed = "None"
     )]
     #[allow(clippy::too_many_arguments)]
     fn new(
-        n_clusters: Option<usize>,
+        n_clusters: usize,
         regr_spec: u8,
         corr_spec: u8,
+        recombination: Recombination,
         kpls_dim: Option<usize>,
-        outdir: Option<String>,
+        _outdir: Option<String>,
         seed: Option<u64>,
     ) -> Self {
         GpMix {
             n_clusters,
             regression_spec: RegressionSpec(regr_spec),
             correlation_spec: CorrelationSpec(corr_spec),
+            recombination,
             kpls_dim,
-            outdir,
+            _outdir,
             seed,
-            xt: None,
-            yt: None,
-            moe_params: None,
+            training_data: None,
+            moe: None,
         }
     }
 
     fn set_training_values(&mut self, xt: PyReadonlyArray2<f64>, yt: PyReadonlyArray2<f64>) {
-        self.xt = Some(xt.as_array().to_owned());
-        self.yt = Some(yt.as_array().to_owned());
+        self.training_data = Some(Dataset::new(
+            xt.as_array().to_owned(),
+            yt.as_array().to_owned(),
+        ));
     }
 
-    fn train(&mut self) {}
-
-    fn predict_values(&self, x: PyReadonlyArray2<f64>) -> PyArray2<f64> {
-        todo!()
+    fn train(&mut self) {
+        let recomb = match self.recombination {
+            Recombination::Hard => egobox_moe::Recombination::Hard,
+            Recombination::Smooth => egobox_moe::Recombination::Smooth(None),
+        };
+        let rng = if let Some(seed) = self.seed {
+            Isaac64Rng::seed_from_u64(seed)
+        } else {
+            Isaac64Rng::from_entropy()
+        };
+        self.moe = Some(
+            Moe::params()
+                .n_clusters(self.n_clusters)
+                .recombination(recomb)
+                .regression_spec(
+                    egobox_moe::RegressionSpec::from_bits(self.regression_spec.0).unwrap(),
+                )
+                .correlation_spec(
[…]
