+++ src/lib.rs
@@ -15,11 +15,11 @@ use pyo3::prelude::*;
 fn egobox(_py: Python, m: &PyModule) -> PyResult<()> {
     pyo3_log::init();
 
+    // utils
     m.add_function(wrap_pyfunction!(to_specs, m)?)?;
-
     m.add_function(wrap_pyfunction!(lhs, m)?)?;
 
-    m.add_class::<Egor>()?;
+    // types
     m.add_class::<RegressionSpec>()?;
     m.add_class::<CorrelationSpec>()?;
     m.add_class::<InfillStrategy>()?;
@@ -29,5 +29,13 @@ fn egobox(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<Vspec>()?;
     m.add_class::<OptimResult>()?;
     m.add_class::<ExpectedOptimum>()?;
+    m.add_class::<Recombination>()?;
+
+    // Surrogate Model
+    m.add_class::<GpMix>()?;
+
+    // Optimizer
+    m.add_class::<Egor>()?;
+
     Ok(())
 }
