### 2022-08-29 #4330 @relf
@arfon Yeah, Rust ML is definitely still in its infancy and Rust is hardly mentioned in your reviewers list. Maybe people from the Rust-ML group who maintain "scikit-learn-for-Rust" project, [linfa](https://github.com/rust-ml/linfa) would be interested in reviewing this software which relies largely on their work, I can mention @bytesnake, @YuhanLiin and @quietlychris.
