### 2022-08-29 #4330 @arfon
:wave: @bytesnake, @YuhanLiin and @quietlychris – would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration here is [_egobox, a Rust toolbox for efficient global optimization_](https://github.com/relf/egobox)
