### 2022-08-29 d05d2ee
Init logging system in moe training method
