### 2022-08-29 4dac853
Add kriging test and example
