+++ python/egobox/tests/test_gpmix.py
@@ -1,41 +1,35 @@
+import unittest
 import numpy as np
-import matplotlib.pyplot as plt
-
 import egobox as egx
+import logging
 
-xt = np.array([[0.0, 1.0, 2.0, 3.0, 4.0]]).T
-yt = np.array([[0.0, 1.0, 1.5, 0.9, 1.0]]).T
+logging.basicConfig(level=logging.INFO)
 
-sm = egx.GpMix(
-    regr_spec=egx.RegressionSpec.CONSTANT,
-    corr_spec=egx.CorrelationSpec.SQUARED_EXPONENTIAL,
-)
-sm.set_training_values(xt, yt)
-sm.train()
 
-num = 100
-x = np.linspace(0.0, 4.0, num).reshape((-1, 1))
+class TestGpMix(unittest.TestCase):
+    def test_gpmix_kriging(self):
+        xt = np.array([[0.0, 1.0, 2.0, 3.0, 4.0]]).T
+        yt = np.array([[0.0, 1.0, 1.5, 0.9, 1.0]]).T
 
-y = sm.predict_values(x)
-# estimated variance
-s2 = sm.predict_variances(x)
+        sm = egx.GpMix(
+            regr_spec=egx.RegressionSpec.CONSTANT,
+            corr_spec=egx.CorrelationSpec.SQUARED_EXPONENTIAL,
+        )
+        sm.set_training_values(xt, yt)
+        sm.train()
 
-fig, axs = plt.subplots(1)
+        # should interpolate
+        self.assertAlmostEqual(1.0, float(sm.predict_values(np.array([[1.0]]))))
+        self.assertAlmostEqual(0.0, float(sm.predict_variances(np.array([[1.0]]))))
 
-# add a plot with variance
-axs.plot(xt, yt, "o")
-axs.plot(x, y)
-axs.fill_between(
-    np.ravel(x),
-    np.ravel(y - 3 * np.sqrt(s2)),
-    np.ravel(y + 3 * np.sqrt(s2)),
-    color="lightgrey",
-)
-axs.set_xlabel("x")
-axs.set_ylabel("y")
-axs.legend(
-    ["Training data", "Prediction", "Confidence Interval 99%"],
-    loc="lower right",
-)
+        # check a point not too far a training point
+        self.assertAlmostEqual(
+            1.1163, float(sm.predict_values(np.array([[1.1]]))), delta=1e-3
+        )
+        self.assertAlmostEqual(
+            0.0, float(sm.predict_variances(np.array([[1.1]]))), delta=1e-3
+        )
 
-plt.show()
+
+if __name__ == "__main__":
+    unittest.main()
