### 2022-08-29 #4330 @bytesnake
Hi @arfon @relf sure I'm happy to help! Just glanced at the documentation (would be first time for me as a JOSS reviewer) and I will run the test/examples, take a look at software quality and paper completeness in the next few days. From the guideline principle section it seems that only a confirmation is needed and no written prose text addressing each point, right? I'm no expert in surrogate models but can comment from Rust perspective :crab:
