### 2022-09-03 b175169
Avoid heaviside optim when one cluster
