+++ moe/src/algorithm.rs
@@ -96,11 +96,10 @@ impl<R: Rng + SeedableRng + Clone> MoeValidParams<f64, R> {
             debug!("Automatic settings {} {:?}", n_clusters, recomb);
         }
 
-        let training = if let Recombination::Smooth(None) = recomb {
+        let training = if recomb == Recombination::Smooth(None) && self.n_clusters() > 1 {
             // Extract 5% of data for validation
             // TODO: Use cross-validation ? Performances
             let (_, training_data) = extract_part(&data, 5);
-
             training_data
         } else {
             data.to_owned()
@@ -136,7 +135,6 @@ impl<R: Rng + SeedableRng + Clone> MoeValidParams<f64, R> {
     ) -> Result<Moe> {
         let gmx = clustering.gmx();
         let recomb = clustering.recombination();
-        let _opt = env_logger::try_init().ok();
         let nx = xt.ncols();
         let data = concatenate(Axis(1), &[xt.view(), yt.view()]).unwrap();
 
@@ -159,7 +157,7 @@ impl<R: Rng + SeedableRng + Clone> MoeValidParams<f64, R> {
             experts.push(expert);
         }
 
-        if recomb == Recombination::Smooth(None) {
+        if recomb == Recombination::Smooth(None) && self.n_clusters() > 1 {
             // Extract 5% of data for validation
             // TODO: Use cross-validation ? Performances
             let (test, _) = extract_part(&data, 5);
@@ -167,6 +165,7 @@ impl<R: Rng + SeedableRng + Clone> MoeValidParams<f64, R> {
             let ytest = Some(test.slice(s![.., nx..]).to_owned());
             let factor =
                 self.optimize_heaviside_factor(&experts, gmx, &xtest.unwrap(), &ytest.unwrap());
+            info!("Retrain mixture with optimized heaviside factor={}", factor);
             let moe = MoeParams::from(self.clone())
                 .n_clusters(gmx.n_clusters())
                 .recombination(Recombination::Smooth(Some(factor)))
