### 2022-09-03 63394b5
Adjust the example to make moe fit as kriging
