### 2022-09-03 1cd49de
Use blas and non-blas version of reduced_likelihood
