+++ gp/src/algorithm.rs
@@ -387,6 +387,7 @@ where
     }
 }
 
+#[cfg(not(feature = "blas"))]
 fn reduced_likelihood<F: Float>(
     fx: &ArrayBase<impl Data<Elem = F>, Ix2>,
     rxx: ArrayBase<impl Data<Elem = F>, Ix2>,
@@ -400,33 +401,17 @@ fn reduced_likelihood<F: Float>(
         r_mx[[ij[0], ij[1]]] = rxx[[i, 0]];
         r_mx[[ij[1], ij[0]]] = rxx[[i, 0]];
     }
-
-    #[cfg(feature = "blas")]
-    let fxl = fx.to_owned().with_lapack();
-    #[cfg(not(feature = "blas"))]
     let fxl = fx;
 
     // R cholesky decomposition
-    #[cfg(feature = "blas")]
-    let r_chol = r_mx.with_lapack().cholesky(UPLO::Lower)?;
-    #[cfg(not(feature = "blas"))]
     let r_chol = r_mx.cholesky()?;
 
     // Solve generalized least squared problem
-    #[cfg(feature = "blas")]
-    let ft = r_chol.solve_triangular(UPLO::Lower, Diag::NonUnit, &fxl)?;
-    #[cfg(not(feature = "blas"))]
     let ft = r_chol.solve_triangular(fxl, UPLO::Lower)?;
 
-    #[cfg(feature = "blas")]
-    let (ft_qr_q, ft_qr_r) = ft.qr().unwrap();
-    #[cfg(not(feature = "blas"))]
     let (ft_qr_q, ft_qr_r) = ft.qr().unwrap().into_decomp();
 
     // Check whether we have an ill-conditionned problem
-    #[cfg(feature = "blas")]
-    let (_, sv_qr_r, _) = ft_qr_r.svd(false, false).unwrap();
-    #[cfg(not(feature = "blas"))]
     let (_, sv_qr_r, _) = ft_qr_r.svd(false, false).unwrap();
     let cond_ft = sv_qr_r[sv_qr_r.len() - 1] / sv_qr_r[0];
     if F::cast(cond_ft) < F::cast(1e-10) {
@@ -445,37 +430,101 @@ fn reduced_likelihood<F: Float>(
             ));
         }
     }
-    #[cfg(feature = "blas")]
-    let yt = r_chol.solve_triangular(
-        UPLO::Lower,
-        Diag::NonUnit,
-        &ytrain.data.to_owned().with_lapack(),
-    )?;
-    #[cfg(not(feature = "blas"))]
     let yt = r_chol.solve_triangular(&ytrain.data, UPLO::Lower)?;
 
-    #[cfg(feature = "blas")]
-    let beta = ft_qr_r.solve_triangular_into(UPLO::Upper, Diag::NonUnit, ft_qr_q.t().dot(&yt))?;
-    #[cfg(not(feature = "blas"))]
     let beta = ft_qr_r.solve_triangular_into(ft_qr_q.t().dot(&yt), UPLO::Upper)?;
 
     let rho = yt - ft.dot(&beta);
     let rho_sqr = rho.mapv(|v| v * v).sum_axis(Axis(0));
-    #[cfg(feature = "blas")]
-    let rho_sqr = rho_sqr.without_lapack();
 
-    #[cfg(feature = "blas")]
-    let gamma = r_chol
-        .t()
-        .solve_triangular_into(UPLO::Upper, Diag::NonUnit, rho)?;
-    #[cfg(not(feature = "blas"))]
     let gamma = r_chol.t().solve_triangular_into(rho, UPLO::Upper)?;
 
     // The determinant of R is equal to the squared product of
[…]
