### 2022-09-03 4ab1e21
Fix blas usage in test and doc
