### 2022-09-04 #4330 @arfon
> Hi @arfon @relf sure I'm happy to help! Just glanced at the documentation (would be first time for me as a JOSS reviewer) and I will run the test/examples, take a look at software quality and paper completeness in the next few days. From the guideline principle section it seems that only a confirmation is needed and no written prose text addressing each point, right? I'm no expert in surrogate models but can comment from Rust perspective 🦀

Thanks @bytesnake. We don't need a detail response letter here but if you find issues associated with the submission, we do ask that you say what they are and ideally open issues on the repository (https://github.com/relf/egobox). You're correct though that there's no 'review letters' being passed around here 😄
