### 2022-09-04 #4737 @editorialbot
[REVIEW]: egobox, a Rust toolbox for efficient global optimization
**Submitting author:** @relf (<a href="http://orcid.org/0000-0001-5479-2961">Rémi Lafage</a>)
**Repository:** https://github.com/relf/egobox
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.5.0
