### 2022-09-07 8b9583b
Clarify BLAS/LAPACK backend doc
