+++ ego/README.md
@@ -24,7 +24,7 @@ It is a Rust port of EGO of the [SMT](https://smt.readthedocs.io) Python library
 There is some usage examples in the examples/ directory. To run, use:
 
 ```
-$ cargo run --release --example ackley --features linfa/intel-mkl-static
+$ cargo run --release --example ackley
 ```
 
 ## License
