+++ README.md
@@ -45,18 +45,16 @@ The `persistent-moe` feature enables `save()` and `load()` methods for MoE model
 
 ### Implementation Notes
 
-`egobox` relies on `[linfa](https://github.com/rust-ml/linfa)` project for methods like clustering and dimension reduction, but also adopts as far as possible
-the same [coding structures](https://github.com/rust-ml/linfa/blob/master/CONTRIBUTE.md).
+`egobox` relies on [linfa](https://github.com/rust-ml/linfa) project for methods like clustering and dimension reduction, but also try to adopt as far as possible
+the same [coding conventions](https://github.com/rust-ml/linfa/blob/master/CONTRIBUTE.md).
 
 Regarding linear algebra routines used in `gp`, `moe` ad `ego`, as for `linfa`, you can:
 
-* either use the pure-Rust implementation [linfa-linalg](https://github.com/rust-ml/linfa-linalg), which is the default 
+* either use the pure-Rust implementation [linfa-linalg](https://github.com/rust-ml/linfa-linalg), which is the default, 
 
 * or you can choose an external BLAS/LAPACK backend available through the [ndarray-linalg](https://github.com/rust-ndarray/ndarray-linalg) project
 
-In this latter case you have to use the `linfa` [BLAS/LAPACK backend features](https://github.com/rust-ml/linfa#blaslapack-backend)
-
-More information in [linfa features](https://github.com/rust-ml/linfa#blaslapack-backend)
+In this latter case you have to use `linfa` [BLAS/LAPACK backend features](https://github.com/rust-ml/linfa#blaslapack-backend) (more information in [linfa features](https://github.com/rust-ml/linfa#blaslapack-backend).
 
 For instance, for using `gp` with the Intel MKL BLAS/Lapack backend, you have to specify in your `Cargo.toml` the following features:
 
@@ -66,7 +64,7 @@ egobox-gp = { version = "0.4.0", features = ["blas", "linfa/intel-mkl-static"] }
 ```
 
 Note: only end-user applications should specify a BLAS/LAPACK provider in `Cargo.toml` (not librairies). 
-In case of a library development, the backend is specified on the command line when building binaries like tests or examples (eg `--features linfa/intel-mkl-static`).
+In case of a library development, the backend is specified on the command line when building binaries like tests or examples (eg `--features blas,linfa/intel-mkl-static`).
 
 ### Examples
 
