### 2022-09-08 036aef5
Remove BLAS/LAPACK ref from module doc
