+++ moe/src/lib.rs
@@ -29,22 +29,6 @@
 //!  
 //! # Features
 //!
-//! ## linfa features
-//!
-//! End-user project should select a BLAS/Lapack backend depending its environment;
-//! it can be either:
-//!
-//! * Openblas: `linfa/openblas-system` or `linfa/openblas-static`
-//! * Netlib: `linfa/netlib-system` or `linfa/netlib-static`
-//! * Intel MKL: `linfa/intel-mkl-system` or `linfa/intel-mkl-static`
-//!
-//! where
-//!
-//! * `*-system` features: try to find the corresponding backend in your installation.
-//! * `*-static` features: try to download and compile the corresponing backend.
-//!
-//! More information in [linfa features](https://github.com/rust-ml/linfa#blaslapack-backend).
-//!
 //! ## persistent
 //!
 //! The `persistent` feature enables `save()`/`load()` methods for a MoE model
