### 2022-09-08 0592eb3
Use ArrayBase when possible and ArrayView otherwise
