### 2022-09-08 b0a8c54
Use ArrayBase in Fit impl for MixintMoeValidParams
