### 2022-09-08 8b0897e
Use ArrayBase in Fit impl for MoeValidParams
