### 2022-09-08 1ac2062
Implement PredictInplace for GP
