+++ gp/src/algorithm.rs
@@ -6,7 +6,7 @@ use crate::utils::{DistanceMatrix, NormalizedMatrix};
 use egobox_doe::{Lhs, SamplingMethod};
 #[cfg(feature = "blas")]
 use linfa::dataset::{WithLapack, WithoutLapack};
-use linfa::prelude::{Dataset, DatasetBase, Fit, Float};
+use linfa::prelude::{Dataset, DatasetBase, Fit, Float, PredictInplace};
 #[cfg(not(feature = "blas"))]
 use linfa_linalg::{cholesky::*, qr::*, svd::*, triangular::*};
 use linfa_pls::PlsRegression;
@@ -196,6 +196,60 @@ impl<F: Float, Mean: RegressionModel<F>, Corr: CorrelationModel<F>> GaussianProc
         } else {
             None
         }
+    }
+}
+
+impl<F, D, Mean, Corr> PredictInplace<ArrayBase<D, Ix2>, Array2<F>>
+    for GaussianProcess<F, Mean, Corr>
+where
+    F: Float,
+    D: Data<Elem = F>,
+    Mean: RegressionModel<F>,
+    Corr: CorrelationModel<F>,
+{
+    fn predict_inplace(&self, x: &ArrayBase<D, Ix2>, y: &mut Array2<F>) {
+        assert_eq!(
+            x.nrows(),
+            y.nrows(),
+            "The number of data points must match the number of output targets."
+        );
+
+        let values = self.predict_values(x).expect("GP Prediction");
+        *y = values;
+    }
+
+    fn default_target(&self, x: &ArrayBase<D, Ix2>) -> Array2<F> {
+        Array2::zeros((x.nrows(), self.ytrain.ncols()))
+    }
+}
+
+struct GpVariancePredictor<'a, F, Mean, Corr>(&'a GaussianProcess<F, Mean, Corr>)
+where
+    F: Float,
+    Mean: RegressionModel<F>,
+    Corr: CorrelationModel<F>;
+
+impl<'a, F, D, Mean, Corr> PredictInplace<ArrayBase<D, Ix2>, Array2<F>>
+    for GpVariancePredictor<'a, F, Mean, Corr>
+where
+    F: Float,
+    D: Data<Elem = F>,
+    Mean: RegressionModel<F>,
+    Corr: CorrelationModel<F>,
+{
+    fn predict_inplace(&self, x: &ArrayBase<D, Ix2>, y: &mut Array2<F>) {
+        assert_eq!(
+            x.nrows(),
+            y.nrows(),
+            "The number of data points must match the number of output targets."
+        );
+
+        let values = self.0.predict_variances(x).expect("GP Prediction");
+        *y = values;
+    }
+
+    fn default_target(&self, x: &ArrayBase<D, Ix2>) -> Array2<F> {
+        Array2::zeros((x.nrows(), self.0.ytrain.ncols()))
     }
 }
 
@@ -556,6 +610,7 @@ mod tests {
     use approx::assert_abs_diff_eq;
     use argmin_testfunctions::rosenbrock;
     use egobox_doe::{Lhs, SamplingMethod};
+    use linfa::prelude::Predict;
     #[cfg(not(feature = "blas"))]
     use linfa_linalg::norm::*;
     use ndarray::{arr2, array, Array, Zip};
@@ -567,7 +622,6 @@ mod tests {
     use paste::paste;
     use rand_isaac::Isaac64Rng;
 
-    // FIXME: Removed nn order to pass with linfa 0.5.1, should be ok with next release
     #[test]
     fn test_constant_function() {
         let dim = 3;
@@ -768,14 +822,17 @@ mod tests {
             Matern52Corr::default(),
         )
[…]
