### 2022-09-09 00f3e1b
Add PredictInplace impl for Moe
