### 2022-09-09 4cb6597
Add PredictInplace for MixintMoe
