### 2022-09-11 56e60f2
Update README (#36)
* Clarify BLAS/LAPACK backend doc

* Fixes

* More README fixes

* Remove BLAS/LAPACK ref from module doc
