### 2022-09-11 5b4e728
Add links to sub-crate in egobox module doc
