### 2022-09-11 4cb1b79
Add contributing doc (#38)
