+++ CONTRIBUTING.md
@@ -1,0 +1,42 @@
+# Contributing
+
+Any input, feedback, bug report or contribution is welcome. 
+
+When contributing to this repository, please first discuss the change you wish to make via an [issue](https://github.com/relf/egobox/issues/new). Subsequent code changes, if any, happen through pull requests.
+
+## Creating a Pull Request
+
+1. Fork the repo and create a branch for your changes from `master`.
+2. If you've added code that should be tested, add tests.
+3. Make sure the style of your code is consistent. Run `cargo fmt` and `cargo clippy`.
+4. If you've changed any functionalities, update the documentation.
+5. Ensure the tests pass. Run `cargo test --release --all`.
+6. Create the pull request.
+
+More information in [GitHub docs](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request). 
+
+## Code of Conduct
+
+Use common sense.
+
+Examples of behavior that contributes to a positive environment for our community include:
+
+* Demonstrating empathy and kindness toward other people
+* Being respectful of differing opinions, viewpoints, and experiences
+* Giving and gracefully accepting constructive feedback
+* Accepting responsibility and apologizing to those affected by our mistakes, and learning from the experience
+* Focusing on what is best not just for us as individuals, but for the overall community
+
+Examples of unacceptable behavior include:
+
+* The use of sexualized language or imagery, and sexual attention or advances of any kind
+* Trolling, insulting or derogatory comments, and personal or political attacks
+* Public or private harassment
+* Publishing others’ private information, such as a physical or email address, without their explicit permission
+* Other conduct which could reasonably be considered inappropriate in a professional setting
+
+(Examples source: [Contributor Covenant](https://www.contributor-covenant.org/version/2/1/code_of_conduct/))
+
+## License
+
+By contributing to this repository, you agree that your contributions will be licensed under its [Apache 2.0 license](https://github.com/relf/egobox/blob/master/LICENSE).
