### 2022-09-11 52b25dc
API improvements  (#37)
* Use ArrayBase when possible and ArrayView otherwise

* Use ArrayBase in Fit impl for MixintMoeValidParams

* Use ArrayBase in Fit impl for MoeValidParams

* Implement PredictInplace for GP

* Add PredictInplace impl for Moe

* Simplify PredictInplace for Moe

* Add PredictInplace for MixintMoe
