### 2022-09-14 38bbe73
Fix error when persistent is enabled
