### 2022-09-20 46f1610
Implement differences
