+++ gp/src/utils.rs
@@ -1,5 +1,5 @@
 use linfa::Float;
-use ndarray::{s, Array1, Array2, ArrayBase, Axis, Data, Ix2};
+use ndarray::{s, Array1, Array2, Array3, ArrayBase, Axis, Data, Ix2};
 #[cfg(feature = "serializable")]
 use serde::{Deserialize, Serialize};
 
@@ -107,10 +107,53 @@ impl<F: Float> DistanceMatrix<F> {
     }
 }
 
+fn differences<F: Float>(
+    x: &ArrayBase<impl Data<Elem = F>, Ix2>,
+    y: &ArrayBase<impl Data<Elem = F>, Ix2>,
+) -> Array3<F> {
+    let x = x.to_owned().insert_axis(Axis(1));
+    let y = y.to_owned().insert_axis(Axis(0));
+    let d = x - y;
+    d
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use approx::assert_abs_diff_eq;
     use ndarray::array;
+
+    #[test]
+    fn test_differences() {
+        let x = array![[-0.9486833], [-0.82219219]];
+        let y = array![
+            [-1.26491106],
+            [-0.63245553],
+            [0.],
+            [0.63245553],
+            [1.26491106]
+        ];
+        assert_abs_diff_eq!(
+            &array![
+                [
+                    [0.31622777],
+                    [-0.31622777],
+                    [-0.9486833],
+                    [-1.58113883],
+                    [-2.21359436]
+                ],
+                [
+                    [0.44271887],
+                    [-0.18973666],
+                    [-0.82219219],
+                    [-1.45464772],
+                    [-2.08710326]
+                ]
+            ],
+            &differences(&x, &y),
+            epsilon = 1e-6
+        )
+    }
 
     #[test]
     fn test_normalized_matrix() {
