### 2022-09-21 78f4896
Improve pairwise differences computation
