### 2022-09-21 6b5e598
Implement predict derivatives for squared exp
