### 2022-09-21 cc421cf
Fix test following cobyla rhobeg adjustment (nlopt intialstep)
