### 2022-09-22 a2ad7a3
Move predict_deriv in specific constant/squaredexp impl
