### 2022-09-22 812435d
Cleanup debug traces
