### 2022-09-23 362aea1
Implement predict_variance_derivatives
