### 2022-09-25 93e3ade
Add predict jacobian  in gp
