+++ moe/src/surrogates.rs
@@ -31,6 +31,11 @@ pub trait Surrogate: std::fmt::Display + Send {
     fn predict_values(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>>;
     /// Predict variance values at n points given as (n, xdim) matrix.
     fn predict_variances(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>>;
+    /// Predict derivatives at n points return (n, xdim) matrix
+    /// where each column is the partial derivatives wrt the ith component
+    fn predict_jacobian(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>>;
+    /// Predict derivatives at n points given as (n, xdim) matrix wrt kx_th component
+    // fn predict_derivatives(&self, x: &ArrayView2<f64>, kx: usize) -> Result<Array2<f64>>;
     /// Save model in given file.
     #[cfg(feature = "persistent")]
     fn save(&self, path: &str) -> Result<()>;
@@ -95,6 +100,9 @@ macro_rules! declare_surrogate {
                 }
                 fn predict_variances(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>> {
                     Ok(self.0.predict_variances(x)?)
+                }
+                fn predict_jacobian(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>> {
+                    Ok(self.0.predict_jacobian(x))
                 }
 
                 #[cfg(feature = "persistent")]
