### 2022-09-25 #4737 @quietlychris
To the best of my ability, I've completed my review of this library. I have tentatively checked the box labeled "Functionality: Have the functional claims of the software been confirmed?"--while I have given the math itself a look-over, as mentioned, I am not an expert in the area. Apart from that caveat, however, I can give a positive review to all other aspects of this submission. If you require anything else from my end in order to move forward, just let me know!
