### 2022-09-26 c176dde
predict variance jacobian hard while smooth is WIP
