+++ moe/src/surrogates.rs
@@ -31,11 +31,12 @@ pub trait Surrogate: std::fmt::Display + Send {
     fn predict_values(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>>;
     /// Predict variance values at n points given as (n, xdim) matrix.
     fn predict_variances(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>>;
-    /// Predict derivatives at n points return (n, xdim) matrix
+    /// Predict derivatives at n points and return (n, xdim) matrix
     /// where each column is the partial derivatives wrt the ith component
     fn predict_jacobian(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>>;
-    /// Predict derivatives at n points given as (n, xdim) matrix wrt kx_th component
-    // fn predict_derivatives(&self, x: &ArrayView2<f64>, kx: usize) -> Result<Array2<f64>>;
+    /// Predict derivatives of the variance at n points and return (n, xdim) matrix
+    /// where each column is the partial derivatives wrt the ith component
+    fn predict_variance_jacobian(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>>;
     /// Save model in given file.
     #[cfg(feature = "persistent")]
     fn save(&self, path: &str) -> Result<()>;
@@ -103,6 +104,9 @@ macro_rules! declare_surrogate {
                 }
                 fn predict_jacobian(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>> {
                     Ok(self.0.predict_jacobian(x))
+                }
+                fn predict_variance_jacobian(&self, x: &ArrayView2<f64>) -> Result<Array2<f64>> {
+                    Ok(self.0.predict_variance_jacobian(x))
                 }
 
                 #[cfg(feature = "persistent")]
