+++ gp/src/algorithm.rs
@@ -10,7 +10,7 @@ use linfa::prelude::{Dataset, DatasetBase, Fit, Float, PredictInplace};
 #[cfg(not(feature = "blas"))]
 use linfa_linalg::{cholesky::*, qr::*, svd::*, triangular::*};
 use linfa_pls::PlsRegression;
-use ndarray::{arr1, s, Array, Array1, Array2, ArrayBase, Axis, Data, Ix1, Ix2, Zip};
+use ndarray::{arr1, s, Array, Array1, Array2, ArrayBase, Axis, Data, Ix2, Zip};
 use ndarray_einsum_beta::*;
 #[cfg(feature = "blas")]
 use ndarray_linalg::{cholesky::*, qr::*, svd::*, triangular::*};
@@ -208,47 +208,67 @@ impl<F: Float, Mean: RegressionModel<F>, Corr: CorrelationModel<F>> GaussianProc
         &self,
         x: &ArrayBase<impl Data<Elem = F>, Ix2>,
         kx: usize,
-    ) -> Array2<F> {
+    ) -> Array1<F> {
         let corr = self._compute_correlation(x);
+        // println!("r = {}", corr);
         let x = (x - &self.xtrain.mean) / &self.xtrain.std;
-        let df = Array2::<F>::zeros((1, x.nrows()));
+
+        let df = Array2::<F>::zeros((1, x.ncols()));
         let beta = &self.inner_params.beta;
         let gamma = &self.inner_params.gamma;
         let df_dx = &df.t().dot(beta);
+        // println!("x = {}", x);
+        // println!("df_dx = {}", df_dx);
+        // println!(
+        //     "shapes df={:?}, beta={:?}, df_dx={:?}",
+        //     df.shape(),
+        //     beta.shape(),
+        //     df_dx.shape()
+        // );
 
         let nr = x.nrows();
         let nc = self.xtrain.data.nrows();
-        let d_dx = &x
+        let d_dx_1 = &x
             .column(kx)
             .to_owned()
             .into_shape((nr, 1))
             .unwrap()
             .broadcast((nr, nc))
             .unwrap()
+            .to_owned();
+
+        let d_dx_2 = self
+            .xtrain
+            .data
+            .column(kx)
             .to_owned()
-            - self
-                .xtrain
-                .data
-                .column(kx)
-                .into_shape((1, nc))
-                .unwrap()
-                .broadcast((nr, nc))
-                .unwrap()
-                .to_owned();
+            .as_standard_layout()
+            .into_shape((1, nc))
+            .unwrap()
+            .to_owned();
 
+        let d_dx = d_dx_1 - d_dx_2;
+        // println!("d_dx = {}", d_dx);
+        // println!("theta = {}", self.theta);
         // Get pairwise componentwise L1-distances to the input training set
-        let theta = &self.theta.to_owned().insert_axis(Axis(0));
-
-        let dcorr = theta.mapv(|v| F::cast(2) * v) * (d_dx * corr);
-        (df_dx - dcorr.dot(gamma)) * &self.ytrain.std / &self.xtrain.std
+        let theta = &self.theta.to_owned();
+        let d_dx_corr = d_dx * corr;
+        let res = (df_dx.row(kx).to_owned()
+            - F::cast(2.) * theta[kx] * d_dx_corr.dot(gamma)[[0, 0]])
+            * self.ytrain.std[0]
+            / self.xtrain.std[kx];
+        res
     }
 
     /// Predict jacobian at one point x
[…]
