### 2022-09-28 f417a07
Add predict_jacobian predict_variance_jacobian
