+++ ego/src/mixint.rs
@@ -452,6 +452,26 @@ impl Surrogate for MixintMoe {
         self.moe.predict_variances(&xcast)
     }
 
+    fn predict_jacobian(&self, x: &ArrayView2<f64>) -> egobox_moe::Result<Array2<f64>> {
+        let mut xcast = if self.work_in_folded_space {
+            unfold_with_enum_mask(&self.xtypes, x)
+        } else {
+            x.to_owned()
+        };
+        cast_to_discrete_values_mut(&self.xtypes, &mut xcast);
+        self.moe.predict_jacobian(&xcast)
+    }
+
+    fn predict_variance_jacobian(&self, x: &ArrayView2<f64>) -> egobox_moe::Result<Array2<f64>> {
+        let mut xcast = if self.work_in_folded_space {
+            unfold_with_enum_mask(&self.xtypes, x)
+        } else {
+            x.to_owned()
+        };
+        cast_to_discrete_values_mut(&self.xtypes, &mut xcast);
+        self.moe.predict_variance_jacobian(&xcast)
+    }
+
     /// Save Moe model in given file.
     #[cfg(feature = "persistent")]
     fn save(&self, path: &str) -> egobox_moe::Result<()> {
