+++ ego/src/utils.rs
@@ -28,10 +28,56 @@ pub fn ei(x: &[f64], obj_model: &dyn ClusteredSurrogate, f_min: f64) -> f64 {
     }
 }
 
+pub fn grad_ei(x: &[f64], obj_model: &dyn ClusteredSurrogate, f_min: f64) -> Array1<f64> {
+    let pt = ArrayView::from_shape((1, x.len()), x).unwrap();
+    if let Ok(p) = obj_model.predict_values(&pt) {
+        if let Ok(s) = obj_model.predict_variances(&pt) {
+            let sigma = s[[0, 0]].sqrt();
+            if sigma.abs() < 1e-12 {
+                Array1::zeros(pt.len())
+            } else {
+                let pred = p[[0, 0]];
+                let diff_y = f_min - pred;
+                let arg = (f_min - pred) / sigma;
+                let y_prime = obj_model.predict_jacobian(&pt).unwrap();
+                let y_prime = y_prime.row(0);
+                let sig_2_prime = obj_model.predict_variance_jacobian(&pt).unwrap();
+
+                let sig_2_prime = sig_2_prime.row(0);
+                let sig_prime = sig_2_prime.mapv(|v| v / (2. * sigma));
+                let arg_prime = y_prime.mapv(|v| v / (-sigma))
+                    - diff_y.to_owned() * sig_prime.mapv(|v| v / (sigma * sigma));
+                let factor = sigma * (-arg / SQRT_2PI) * (-(arg * arg) / 2.).exp();
+
+                let arg1 = y_prime.mapv(|v| v * (-norm_cdf(arg)));
+                let arg2 = diff_y * norm_pdf(arg) * arg_prime.to_owned();
+                let arg3 = sig_prime.to_owned() * norm_pdf(arg);
+                let arg4 = factor * arg_prime;
+                arg1 + arg2 + arg3 + arg4
+            }
+        } else {
+            Array1::zeros(pt.len())
+        }
+    } else {
+        Array1::zeros(pt.len())
+    }
+}
+
 pub fn wb2s(x: &[f64], obj_model: &dyn ClusteredSurrogate, f_min: f64, scale: f64) -> f64 {
     let pt = ArrayView::from_shape((1, x.len()), x).unwrap();
     let ei = ei(x, obj_model, f_min);
     scale * ei - obj_model.predict_values(&pt).unwrap()[[0, 0]]
+}
+
+pub fn grad_wbs2(
+    x: &[f64],
+    obj_model: &dyn ClusteredSurrogate,
+    f_min: f64,
+    scale: f64,
+) -> Array1<f64> {
+    let pt = ArrayView::from_shape((1, x.len()), x).unwrap();
+    let grad_ei = grad_ei(x, obj_model, f_min) * scale;
+    grad_ei - obj_model.predict_jacobian(&pt).unwrap().row(0)
 }
 
 pub fn compute_wb2s_scale(
@@ -125,7 +171,10 @@ pub fn update_data(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use ndarray::array;
+    use egobox_doe::SamplingMethod;
+    use egobox_moe::*;
+    use linfa::prelude::*;
+    use ndarray::{array, Array};
 
     #[test]
     fn test_is_update_ok() {
@@ -149,5 +198,44 @@ mod tests {
         );
         assert_eq!(&array![[0., 1.], [2., 3.], [3., 4.]], xdata);
         assert_eq!(&array![[3.], [4.], [6.]], ydata);
+    }
+
+    fn sphere(x: &Array2<f64>) -> Array2<f64> {
+        let s = (x * x).sum_axis(Axis(1));
[…]
