### 2022-10-04 #4737 @bytesnake
Finished my review to the best of my ability. The code quality adheres high standard, is automated tested and fully documented. It fulfills the promises by providing some useful examples. The paper is well-written (there is one "gausian", spot single "s", typo), has a good introduction and a clear statement of need. Finally references to papers and established software packages are in place and complete.
