### 2022-10-05 45e5a6d
Fix examples doc section
