### 2022-10-07 d31e12c
JOSS paper  (#42)
* Initial paper version

* Adjust SMT reference

* Fix affiliation

* Fix DOIs

* Review statement of need

* Add architecture figure

* Fix typo

* Fix examples doc section

* Absolute links for sub-packages

* Re-use README as crate doc

* Absolute link for tutorial
