+++ joss/paper.bib
@@ -1,0 +1,77 @@
+@article{Jones1998,
+  title={Efficient global optimization of expensive black-box functions},
+  author={Jones, Donald R and Schonlau, Matthias and Welch, William J},
+  journal={Journal of Global optimization},
+  volume={13},
+  number={4},
+  pages={455--492},
+  year={1998},
+  publisher={Springer}
+}
+
+@article{Bouhlel2016,
+  title={Improving kriging surrogates of high-dimensional design models by Partial Least Squares dimension reduction},
+  author={Bouhlel, Mohamed Amine and Bartoli, Nathalie and Otsmane, Abdelkader and Morlier, Joseph},
+  journal={Structural and Multidisciplinary Optimization},
+  volume={53},
+  number={5},
+  pages={935--952},
+  doi={10.1007/s00158-015-1395-9},
+  year={2016},
+  publisher={Springer},
+}
+
+@inproceedings{Bartoli2016,
+  title={Improvement of efficient global optimization with application to aircraft wing design},
+  author={Bartoli, Nathalie and Bouhlel, Mohamed-Amine and Kurek, Igor and Lafage, R{\'e}mi and Lefebvre, Thierry and Morlier, Joseph and Priem, R{\'e}my and Stilz, Vivien and Regis, Rommel},
+  booktitle={17th AIAA/ISSMO Multidisciplinary analysis and optimization conference},
+  pages={4001},
+  doi={10.2514/6.2016-4001},
+  year={2016},
+}
+
+@article{Bartoli2019,
+  title = {Adaptive modeling strategy for constrained global optimization with application to aerodynamic wing design},
+  journal = {Aerospace Science and Technology},
+  volume = {90},
+  pages = {85-102},
+  doi = {10.1016/j.ast.2019.03.041},
+  year = {2019},
+  issn = {1270-9638},
+  url = {https://www.sciencedirect.com/science/article/pii/S1270963818306011},
+  author = {N. Bartoli and T. Lefebvre and S. Dubreuil and R. Olivanti and R. Priem and N. Bons and J.R.R.A. Martins and J. Morlier},
+  keywords = {Surrogate modeling, Global optimization, Multimodal optimization, Mixture of experts, Aerodynamic shape optimization, Wing design},
+  abstract = {Surrogate models are often used to reduce the cost of design optimization problems that involve computationally costly models, such as computational fluid dynamics simulations. However, the number of evaluations required by surrogate models usually scales poorly with the number of design variables, and there is a need for both better constraint formulations and multimodal function handling. To address this issue, we developed a surrogate-based gradient-free optimization algorithm that can handle cases where the function evaluations are expensive, the computational budget is limited, the functions are multimodal, and the optimization problem includes nonlinear equality or inequality constraints. The proposed algorithm—super efficient global optimization coupled with mixture of experts (SEGOMOE)—can tackle complex constrained design optimization problems through the use of an enrichment strategy based on a mixture of experts coupled with adaptive surrogate models. The performance of this approach was evaluated for analytic constrained and unconstrained problems, as well as for a multimodal aerodynamic shape optimization problem with 17 design variables and an equality constraint. Our results showed that the method is efficient and that the optimum is much less dependent on the starting point than the conventional gradient-based optimization.}
+}
+
+@article{Dubreuil2020,
+  title={Towards an efficient global multidisciplinary design optimization algorithm},
+  author={Dubreuil, Sylvain and Bartoli, Nathalie and Gogu, Christian and Lefebvre, Thierry},
[…]
