+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-Version 0.5.0 - unreleased
+Version 0.6.0 - unreleased
 ==========================
 
 * Add `GpMix`/`Gpx` in Python `egobox` module wrapping `egobox-moe::Moe` (#31)
+
+Version 0.5.0 - 2022-10-07
+==========================
+
 * Add Egor `minimize` interruption capability (Ctrl+C) from Python (#30)
 * Minor performance improvement in moe clustering (#29)
-* Improvements following JOSS review (#34, #36, #38, #39, #40)
+* Improvements following JOSS submission review (#34, #36, #38, #39, #40, #42)
 
 Changes
 -------
