### 2022-10-07 7f1e1d9
Release 0.5.0 (#32)
* Update changelog

* Use central difference for models derivatives

* Update changelog

* Bump 0.5.0

* Update CHANGELOG
