+++ pyproject.toml
@@ -23,7 +23,7 @@ requires = ["maturin>=0.12,<0.13", "poetry_core>=1.0.0"]
 
 [tool.poetry]
 name = "egobox"
-version = "0.4.0"
+version = "0.5.0"
 description = "Python binding for egobox EGO optimizer written in Rust"
 authors = ["Rémi Lafage <remi.lafage@onera.fr>"]
 
