### 2022-10-08 #4737 @arfon
@relf – could you please merge this PR with some small changes to your paper? https://github.com/relf/egobox/pull/43
