### 2022-10-08 2039e8b
Update paper.md (#43)
