### 2022-10-09 #4737 @editorialbot
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
